% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzTabWriter.r
\name{as.data.frame}
\alias{as.data.frame}
\title{Converts the provided Metadata object to a ragged data frame.}
\usage{
as.data.frame(metadata)
}
\arguments{
\item{metadata}{the R6 Metadata object to write.}
}
\description{
Converts the provided Metadata object to a ragged data frame.
}
