% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{as.scalar}
\alias{as.scalar}
\title{Extends jsonlite's as.scalar method.}
\usage{
as.scalar(obj)
}
\arguments{
\item{obj}{The object to convert to a scalar}
}
\description{
This function is originally from the jsonlite package, file as.scalar.R
Unfortunately, that function is not exported from jsonlite, but we need it for 
our more safe_unbox function above. The call to is.namedlist has been 
inlined with the actual code.
}
