% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\docType{class}
\name{Assay}
\alias{Assay}
\title{Assay}
\format{
An \code{R6Class} generator object
}
\description{
Assay Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{name}}{character}

\item{\code{custom}}{list( \link{Parameter} ) [optional]}

\item{\code{external_uri}}{character [optional]}

\item{\code{sample_ref}}{\link{Sample} [optional]}

\item{\code{ms_run_ref}}{list( \link{MsRun} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Assay-new}{\code{Assay$new()}}
\item \href{#method-Assay-toJSON}{\code{Assay$toJSON()}}
\item \href{#method-Assay-fromJSON}{\code{Assay$fromJSON()}}
\item \href{#method-Assay-toJSONString}{\code{Assay$toJSONString()}}
\item \href{#method-Assay-fromJSONString}{\code{Assay$fromJSONString()}}
\item \href{#method-Assay-toDataFrame}{\code{Assay$toDataFrame()}}
\item \href{#method-Assay-fromDataFrame}{\code{Assay$fromDataFrame()}}
\item \href{#method-Assay-clone}{\code{Assay$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-new"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-new}{}}}
\subsection{Method \code{new()}}{
Create an Assay
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$new(
  name,
  ms_run_ref,
  id = NULL,
  custom = NULL,
  external_uri = NULL,
  sample_ref = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Assay name.}

\item{\code{ms_run_ref}}{Reference to \link{MsRun}s.}

\item{\code{id}}{Assay id.}

\item{\code{custom}}{Custom parameters.}

\item{\code{external_uri}}{External URI.}

\item{\code{sample_ref}}{Reference to \link{Sample}.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$fromJSON(AssayJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AssayJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$fromJSONString(AssayJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AssayJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from assay data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$fromDataFrame(AssayDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AssayDataFrame}}{Assay data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
