% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_all_of.R
\docType{class}
\name{AssayAllOf}
\alias{AssayAllOf}
\title{AssayAllOf}
\format{An \code{R6Class} generator object}
\description{
AssayAllOf Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character}

\item{\code{custom}}{list( \link{Parameter} ) [optional]}

\item{\code{external_uri}}{character [optional]}

\item{\code{sample_ref}}{\link{Sample} [optional]}

\item{\code{ms_run_ref}}{list( \link{MsRun} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AssayAllOf$new()}}
\item \href{#method-toJSON}{\code{AssayAllOf$toJSON()}}
\item \href{#method-fromJSON}{\code{AssayAllOf$fromJSON()}}
\item \href{#method-toJSONString}{\code{AssayAllOf$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AssayAllOf$fromJSONString()}}
\item \href{#method-clone}{\code{AssayAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayAllOf$new(
  name,
  ms_run_ref,
  custom = NULL,
  external_uri = NULL,
  sample_ref = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayAllOf$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayAllOf$fromJSON(AssayAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayAllOf$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayAllOf$fromJSONString(AssayAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssayAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
