% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\docType{class}
\name{Comment}
\alias{Comment}
\title{Comment}
\format{
An \code{R6Class} generator object
}
\description{
Comment Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prefix}}{character}

\item{\code{msg}}{character}

\item{\code{line_number}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Comment-new}{\code{Comment$new()}}
\item \href{#method-Comment-toJSON}{\code{Comment$toJSON()}}
\item \href{#method-Comment-fromJSON}{\code{Comment$fromJSON()}}
\item \href{#method-Comment-toJSONString}{\code{Comment$toJSONString()}}
\item \href{#method-Comment-fromJSONString}{\code{Comment$fromJSONString()}}
\item \href{#method-Comment-toDataFrame}{\code{Comment$toDataFrame()}}
\item \href{#method-Comment-fromDataFrame}{\code{Comment$fromDataFrame()}}
\item \href{#method-Comment-clone}{\code{Comment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-new"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-new}{}}}
\subsection{Method \code{new()}}{
Create a Comment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$new(prefix, msg, line_number = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prefix}}{'COM'.}

\item{\code{msg}}{Comment content.}

\item{\code{line_number}}{Line number in the mzTab-M file.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$fromJSON(CommentJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CommentJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$fromJSONString(CommentJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CommentJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from comment data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$fromDataFrame(CommentDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CommentDataFrame}}{Comment data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
