% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\docType{class}
\name{Contact}
\alias{Contact}
\title{Contact}
\format{
An \code{R6Class} generator object
}
\description{
Contact Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{name}}{character [optional]}

\item{\code{affiliation}}{character [optional]}

\item{\code{email}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Contact-new}{\code{Contact$new()}}
\item \href{#method-Contact-toJSON}{\code{Contact$toJSON()}}
\item \href{#method-Contact-fromJSON}{\code{Contact$fromJSON()}}
\item \href{#method-Contact-toJSONString}{\code{Contact$toJSONString()}}
\item \href{#method-Contact-fromJSONString}{\code{Contact$fromJSONString()}}
\item \href{#method-Contact-toDataFrame}{\code{Contact$toDataFrame()}}
\item \href{#method-Contact-fromDataFrame}{\code{Contact$fromDataFrame()}}
\item \href{#method-Contact-clone}{\code{Contact$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Contact-new"></a>}}
\if{latex}{\out{\hypertarget{method-Contact-new}{}}}
\subsection{Method \code{new()}}{
Create a Contact
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contact$new(id = NULL, name = NULL, affiliation = NULL, email = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Contact id.}

\item{\code{name}}{Contact name.}

\item{\code{affiliation}}{Contact affiliation / address.}

\item{\code{email}}{Contact email.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Contact-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Contact-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contact$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Contact-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Contact-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contact$fromJSON(ContactJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ContactJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Contact-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Contact-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contact$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Contact-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Contact-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contact$fromJSONString(ContactJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ContactJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Contact-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Contact-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contact$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Contact-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Contact-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from contact data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contact$fromDataFrame(ContactDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ContactDataFrame}}{contact data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Contact-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Contact-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Contact$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
