% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\docType{class}
\name{CV}
\alias{CV}
\title{CV}
\format{
An \code{R6Class} generator object
}
\description{
CV Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{label}}{character}

\item{\code{full_name}}{character}

\item{\code{version}}{character}

\item{\code{uri}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CV-new}{\code{CV$new()}}
\item \href{#method-CV-toJSON}{\code{CV$toJSON()}}
\item \href{#method-CV-fromJSON}{\code{CV$fromJSON()}}
\item \href{#method-CV-toJSONString}{\code{CV$toJSONString()}}
\item \href{#method-CV-fromJSONString}{\code{CV$fromJSONString()}}
\item \href{#method-CV-toDataFrame}{\code{CV$toDataFrame()}}
\item \href{#method-CV-fromDataFrame}{\code{CV$fromDataFrame()}}
\item \href{#method-CV-clone}{\code{CV$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CV-new"></a>}}
\if{latex}{\out{\hypertarget{method-CV-new}{}}}
\subsection{Method \code{new()}}{
Create a CV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$new(label, full_name, version, uri, id = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{CV label.}

\item{\code{full_name}}{CV full name.}

\item{\code{version}}{CV version.}

\item{\code{uri}}{URI to external CV.}

\item{\code{id}}{URI id.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CV-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CV-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CV-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CV-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$fromJSON(CVJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CVJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CV-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CV-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CV-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CV-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$fromJSONString(CVJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CVJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CV-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-CV-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CV-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-CV-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from CV data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$fromDataFrame(CvDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CvDataFrame}}{CV data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CV-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CV-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
