% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_all_of.R
\docType{class}
\name{DatabaseAllOf}
\alias{DatabaseAllOf}
\title{DatabaseAllOf}
\format{An \code{R6Class} generator object}
\description{
DatabaseAllOf Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{param}}{\link{Parameter}}

\item{\code{prefix}}{character}

\item{\code{version}}{character}

\item{\code{uri}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatabaseAllOf$new()}}
\item \href{#method-toJSON}{\code{DatabaseAllOf$toJSON()}}
\item \href{#method-fromJSON}{\code{DatabaseAllOf$fromJSON()}}
\item \href{#method-toJSONString}{\code{DatabaseAllOf$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DatabaseAllOf$fromJSONString()}}
\item \href{#method-clone}{\code{DatabaseAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseAllOf$new(param, prefix, version, uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseAllOf$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseAllOf$fromJSON(DatabaseAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseAllOf$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseAllOf$fromJSONString(DatabaseAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
