% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mz_tab.R
\name{extractComments}
\alias{extractComments}
\title{Extract the Comment from data frame.}
\usage{
extractComments(mztab.table)
}
\arguments{
\item{mztab.table}{the mztab data frame}
}
\description{
Extract the \link{Comment}s from an mztab data frame.
}
