% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzTabReader.r
\name{extractMetadata}
\alias{extractMetadata}
\title{Read the metadata from an mztab table and return the Metadata object.}
\usage{
extractMetadata(mztab.table)
}
\arguments{
\item{the}{mztab table}
}
\description{
Read the metadata from an mztab table and return the Metadata object.
}
