% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mz_tab.R
\name{extractSmallMoleculeEvidence}
\alias{extractSmallMoleculeEvidence}
\title{Extract the SmallMoleculeEvidence from data frame.}
\usage{
extractSmallMoleculeEvidence(mztab.table)
}
\arguments{
\item{mztab.table}{the mztab data frame}
}
\description{
Extract the \link{SmallMoleculeEvidence} from an mztab data frame.
}
