% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mz_tab.R
\name{extractSmallMoleculeFeatures}
\alias{extractSmallMoleculeFeatures}
\title{Extract the SmallMoleculeFeatures from data frame.}
\usage{
extractSmallMoleculeFeatures(mztab.table)
}
\arguments{
\item{mztab.table}{the mztab data frame}
}
\description{
Extract the \link{SmallMoleculeFeature} from an mztab data frame.
}
