% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mz_tab.R
\name{extractSmallMoleculeSummary}
\alias{extractSmallMoleculeSummary}
\title{Extract the SmallMoleculeSummary from data frame.}
\usage{
extractSmallMoleculeSummary(mztab.table)
}
\arguments{
\item{mztab.table}{the mztab data frame}
}
\description{
Extract the \link{SmallMoleculeSummary} from an mztab data frame.
}
