% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument.R
\docType{class}
\name{Instrument}
\alias{Instrument}
\title{Instrument}
\format{
An \code{R6Class} generator object
}
\description{
Instrument Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{name}}{\link{Parameter} [optional]}

\item{\code{source}}{\link{Parameter} [optional]}

\item{\code{analyzer}}{list( \link{Parameter} ) [optional]}

\item{\code{detector}}{\link{Parameter} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Instrument-new}{\code{Instrument$new()}}
\item \href{#method-Instrument-toJSON}{\code{Instrument$toJSON()}}
\item \href{#method-Instrument-fromJSON}{\code{Instrument$fromJSON()}}
\item \href{#method-Instrument-toJSONString}{\code{Instrument$toJSONString()}}
\item \href{#method-Instrument-fromJSONString}{\code{Instrument$fromJSONString()}}
\item \href{#method-Instrument-toDataFrame}{\code{Instrument$toDataFrame()}}
\item \href{#method-Instrument-fromDataFrame}{\code{Instrument$fromDataFrame()}}
\item \href{#method-Instrument-clone}{\code{Instrument$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-new"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-new}{}}}
\subsection{Method \code{new()}}{
Create an Instrument
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$new(
  id = NULL,
  name = NULL,
  source = NULL,
  analyzer = NULL,
  detector = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Instrument id.}

\item{\code{name}}{Instrument name \link{Parameter}.}

\item{\code{source}}{Instrument source \link{Parameter}.}

\item{\code{analyzer}}{Instrument analyzer list( \link{Parameter} )s.}

\item{\code{detector}}{Instrument detector \link{Parameter}.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$fromJSON(InstrumentJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{InstrumentJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$fromJSONString(InstrumentJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{InstrumentJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from instrument data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$fromDataFrame(InstrumentDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{InstrumentDataFrame}}{Instrument data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instrument-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Instrument-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instrument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
