% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_all_of.R
\docType{class}
\name{InstrumentAllOf}
\alias{InstrumentAllOf}
\title{InstrumentAllOf}
\format{An \code{R6Class} generator object}
\description{
InstrumentAllOf Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{\link{Parameter} [optional]}

\item{\code{source}}{\link{Parameter} [optional]}

\item{\code{analyzer}}{list( \link{Parameter} ) [optional]}

\item{\code{detector}}{\link{Parameter} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InstrumentAllOf$new()}}
\item \href{#method-toJSON}{\code{InstrumentAllOf$toJSON()}}
\item \href{#method-fromJSON}{\code{InstrumentAllOf$fromJSON()}}
\item \href{#method-toJSONString}{\code{InstrumentAllOf$toJSONString()}}
\item \href{#method-fromJSONString}{\code{InstrumentAllOf$fromJSONString()}}
\item \href{#method-clone}{\code{InstrumentAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentAllOf$new(
  name = NULL,
  source = NULL,
  analyzer = NULL,
  detector = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentAllOf$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentAllOf$fromJSON(InstrumentAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentAllOf$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentAllOf$fromJSONString(InstrumentAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstrumentAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
