% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_run_all_of.R
\docType{class}
\name{MsRunAllOf}
\alias{MsRunAllOf}
\title{MsRunAllOf}
\format{An \code{R6Class} generator object}
\description{
MsRunAllOf Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character}

\item{\code{location}}{character}

\item{\code{instrument_ref}}{\link{Instrument} [optional]}

\item{\code{format}}{\link{Parameter} [optional]}

\item{\code{id_format}}{\link{Parameter} [optional]}

\item{\code{fragmentation_method}}{list( \link{Parameter} ) [optional]}

\item{\code{scan_polarity}}{list( \link{Parameter} ) [optional]}

\item{\code{hash}}{character [optional]}

\item{\code{hash_method}}{\link{Parameter} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MsRunAllOf$new()}}
\item \href{#method-toJSON}{\code{MsRunAllOf$toJSON()}}
\item \href{#method-fromJSON}{\code{MsRunAllOf$fromJSON()}}
\item \href{#method-toJSONString}{\code{MsRunAllOf$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MsRunAllOf$fromJSONString()}}
\item \href{#method-clone}{\code{MsRunAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRunAllOf$new(
  name,
  location,
  instrument_ref = NULL,
  format = NULL,
  id_format = NULL,
  fragmentation_method = NULL,
  scan_polarity = NULL,
  hash = NULL,
  hash_method = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRunAllOf$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRunAllOf$fromJSON(MsRunAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRunAllOf$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRunAllOf$fromJSONString(MsRunAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRunAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
