% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTD.R
\name{MTD-export}
\alias{MTD-export}
\title{Defining and exporting the mzTab-M metadata table}
\description{
The metadata section/table of the mzTab-M definition is comprehensive, but
also tricky to define. The *rmzTabM* package provides a variety of utility
functions that help defining this information. These might be re-used for
software package developers to export metabolomics results from their
respective software. Importantly, the helper functions listed here only
define the core elements for the MTD section, helping with re-arranging and
reformatting information available e.g. in `data.frame` format into the
respective fields in the MTD section. Additional (optional) fields might
need to be added manually depending on availability for an experiment.

See also the [specification of the MTD section](https://github.com/HUPO-PSI/mzTab-M/blob/main/specification_documents/mzTab_format_specification_2_1-M.adoc#62-metadata-section)
for details and more information.

Generally, MTD data can be categarized into the following parts:

- *Core information*: general information on the experiment. A minimal
  set can be created using the [mtb_skeleton()] function, which might be
  further expanded with additional fields.

- *Sample information*:

- *MS run information*: information on the individual MS *runs*
  (measurements of the samples). Each data file is one run. Use the
  [mtd_ms_run()] function to define this part of the metadata section.

- *Assay information*:

- *Study variable information*:


The helper function listed above can be used sequentially to create the
metadata information. See the examples below for a general approach how to
define the MTD section of an experiment.

In addition, various helper functions are available to assist in MTD data
generation:

- [mtd_sort()]: to sort the MTD `matrix` into the expected order.
- [mtd_fields()]: helps formatting values into the mzTab-M-specific format.
}
\author{
Johannes Rainer, Philippine Louail
}
