% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTD.R
\name{mtd_sort}
\alias{mtd_sort}
\title{Sort rows in a MTD matrix to match the expected order}
\usage{
mtd_sort(x)
}
\arguments{
\item{x}{two-column matrix with the first column containing the metadata
field names.}
}
\value{
input parameter `x` sorted into the correct order.
}
\description{
Helper function to sort a mzTab-M *MTD* `matrix`, such as generated by
[mtd_skeleton()], into the correct order of the metadata fields.
}
\author{
Johannes Rainer
}
