% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_tab.R
\docType{class}
\name{MzTab}
\alias{MzTab}
\title{MzTab}
\format{
An \code{R6Class} generator object
}
\description{
MzTab Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{\link{Metadata}}

\item{\code{smallMoleculeSummary}}{list( \link{SmallMoleculeSummary} )}

\item{\code{smallMoleculeFeature}}{list( \link{SmallMoleculeFeature} )}

\item{\code{smallMoleculeEvidence}}{list( \link{SmallMoleculeEvidence} )}

\item{\code{comment}}{list( \link{Comment} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MzTab-new}{\code{MzTab$new()}}
\item \href{#method-MzTab-toJSON}{\code{MzTab$toJSON()}}
\item \href{#method-MzTab-fromJSON}{\code{MzTab$fromJSON()}}
\item \href{#method-MzTab-toJSONString}{\code{MzTab$toJSONString()}}
\item \href{#method-MzTab-fromJSONString}{\code{MzTab$fromJSONString()}}
\item \href{#method-MzTab-toDataFrame}{\code{MzTab$toDataFrame()}}
\item \href{#method-MzTab-fromDataFrame}{\code{MzTab$fromDataFrame()}}
\item \href{#method-MzTab-clone}{\code{MzTab$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzTab-new"></a>}}
\if{latex}{\out{\hypertarget{method-MzTab-new}{}}}
\subsection{Method \code{new()}}{
Create an MzTab object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzTab$new(
  metadata,
  smallMoleculeSummary,
  smallMoleculeFeature,
  smallMoleculeEvidence,
  comment = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{The metadata \link{Metadata}.}

\item{\code{smallMoleculeSummary}}{The list( \link{SmallMoleculeSummary} ).}

\item{\code{smallMoleculeFeature}}{The list( \link{SmallMoleculeFeature} ).}

\item{\code{smallMoleculeEvidence}}{The list( \link{SmallMoleculeEvidence} ).}

\item{\code{comment}}{The list( \link{Comment} ).}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzTab-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MzTab-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzTab$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzTab-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MzTab-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzTab$fromJSON(MzTabJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MzTabJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzTab-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MzTab-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzTab$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzTab-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MzTab-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzTab$fromJSONString(MzTabJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MzTabJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzTab-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-MzTab-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzTab$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzTab-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-MzTab-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from MzTab data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzTab$fromDataFrame(MzTabDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MzTabDataFrame}}{MzTab data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MzTab-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MzTab-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MzTab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
