% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_column_mapping.R
\docType{class}
\name{OptColumnMapping}
\alias{OptColumnMapping}
\title{OptColumnMapping}
\format{
An \code{R6Class} generator object
}
\description{
OptColumnMapping Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{identifier}}{character}

\item{\code{param}}{\link{Parameter} [optional]}

\item{\code{value}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptColumnMapping-new}{\code{OptColumnMapping$new()}}
\item \href{#method-OptColumnMapping-toJSON}{\code{OptColumnMapping$toJSON()}}
\item \href{#method-OptColumnMapping-fromJSON}{\code{OptColumnMapping$fromJSON()}}
\item \href{#method-OptColumnMapping-toJSONString}{\code{OptColumnMapping$toJSONString()}}
\item \href{#method-OptColumnMapping-fromJSONString}{\code{OptColumnMapping$fromJSONString()}}
\item \href{#method-OptColumnMapping-toString}{\code{OptColumnMapping$toString()}}
\item \href{#method-OptColumnMapping-fromString}{\code{OptColumnMapping$fromString()}}
\item \href{#method-OptColumnMapping-clone}{\code{OptColumnMapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptColumnMapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptColumnMapping-new}{}}}
\subsection{Method \code{new()}}{
Create an OptColumnMapping
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptColumnMapping$new(identifier, param = NULL, value = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifier}}{The Identifier for the mapping.}

\item{\code{param}}{The CV \link{Parameter} defining the type of the value.}

\item{\code{value}}{The value.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptColumnMapping-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OptColumnMapping-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptColumnMapping$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptColumnMapping-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OptColumnMapping-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptColumnMapping$fromJSON(OptColumnMappingJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OptColumnMappingJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptColumnMapping-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OptColumnMapping-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptColumnMapping$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptColumnMapping-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OptColumnMapping-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptColumnMapping$fromJSONString(OptColumnMappingJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OptColumnMappingJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptColumnMapping-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OptColumnMapping-toString}{}}}
\subsection{Method \code{toString()}}{
Serialize to string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptColumnMapping$toString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptColumnMapping-fromString"></a>}}
\if{latex}{\out{\hypertarget{method-OptColumnMapping-fromString}{}}}
\subsection{Method \code{fromString()}}{
Deserialize from string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptColumnMapping$fromString(OptColumnMappingString)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OptColumnMappingString}}{The string representation to deserialize}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptColumnMapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptColumnMapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptColumnMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
