% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\docType{class}
\name{Parameter}
\alias{Parameter}
\title{Parameter}
\format{
An \code{R6Class} generator object
}
\description{
Parameter Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{cv_label}}{character [optional]}

\item{\code{cv_accession}}{character [optional]}

\item{\code{name}}{character}

\item{\code{value}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Parameter-new}{\code{Parameter$new()}}
\item \href{#method-Parameter-toJSON}{\code{Parameter$toJSON()}}
\item \href{#method-Parameter-fromJSON}{\code{Parameter$fromJSON()}}
\item \href{#method-Parameter-toJSONString}{\code{Parameter$toJSONString()}}
\item \href{#method-Parameter-fromJSONString}{\code{Parameter$fromJSONString()}}
\item \href{#method-Parameter-toString}{\code{Parameter$toString()}}
\item \href{#method-Parameter-fromString}{\code{Parameter$fromString()}}
\item \href{#method-Parameter-clone}{\code{Parameter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-new}{}}}
\subsection{Method \code{new()}}{
Create a Parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$new(name, value, id = NULL, cv_label = "", cv_accession = "", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Parameter name.}

\item{\code{value}}{Parameter value.}

\item{\code{id}}{Parameter id.}

\item{\code{cv_label}}{The label of the controlled vocabulary.}

\item{\code{cv_accession}}{The accession number of this Parameter.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$fromJSON(ParameterJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ParameterJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$fromJSONString(ParameterJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ParameterJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-toString}{}}}
\subsection{Method \code{toString()}}{
Serialize to string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$toString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-fromString"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-fromString}{}}}
\subsection{Method \code{fromString()}}{
Deserialize from string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$fromString(id, ParameterString)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Optional id of the Parameter.}

\item{\code{ParameterString}}{String representation of the Parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parameter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Parameter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
