% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzTabReader.r
\name{readMzTab}
\alias{readMzTab}
\title{Read an mzTab tab separated file from the passed in file.}
\usage{
readMzTab(filename)
}
\arguments{
\item{filename}{the mzTab file to parse.}
}
\description{
Read an mzTab tab separated file from the passed in file.
}
