% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mz_tab.R
\name{readMzTabString}
\alias{readMzTabString}
\title{Read an mzTab-M string into a data frame.}
\usage{
readMzTabString(mzTabString)
}
\arguments{
\item{mzTabString}{the mzTab string to parse.}
}
\description{
Read an mzTab tab separated file from the passed in string.
This file creates a temporary file on disk to check for the maximum number of columns.
The returned object is a data.frame as returned by utils::read.table.
}
