% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\docType{class}
\name{SampleProcessing}
\alias{SampleProcessing}
\title{SampleProcessing}
\format{
An \code{R6Class} generator object
}
\description{
SampleProcessing Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{sampleProcessing}}{list( \link{Parameter} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SampleProcessing-new}{\code{SampleProcessing$new()}}
\item \href{#method-SampleProcessing-toJSON}{\code{SampleProcessing$toJSON()}}
\item \href{#method-SampleProcessing-fromJSON}{\code{SampleProcessing$fromJSON()}}
\item \href{#method-SampleProcessing-toJSONString}{\code{SampleProcessing$toJSONString()}}
\item \href{#method-SampleProcessing-fromJSONString}{\code{SampleProcessing$fromJSONString()}}
\item \href{#method-SampleProcessing-toDataFrame}{\code{SampleProcessing$toDataFrame()}}
\item \href{#method-SampleProcessing-fromDataFrame}{\code{SampleProcessing$fromDataFrame()}}
\item \href{#method-SampleProcessing-clone}{\code{SampleProcessing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleProcessing-new"></a>}}
\if{latex}{\out{\hypertarget{method-SampleProcessing-new}{}}}
\subsection{Method \code{new()}}{
Create a SampleProcessing object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$new(id = NULL, sampleProcessing = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{SampleProcessing id.}

\item{\code{sampleProcessing}}{Sample processing list( \link{Parameter} ).}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleProcessing-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SampleProcessing-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleProcessing-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SampleProcessing-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$fromJSON(SampleProcessingJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SampleProcessingJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleProcessing-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SampleProcessing-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleProcessing-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SampleProcessing-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$fromJSONString(SampleProcessingJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SampleProcessingJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleProcessing-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-SampleProcessing-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleProcessing-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-SampleProcessing-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from SampleProcessingDataFrame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$fromDataFrame(SampleProcessingDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SampleProcessingDataFrame}}{Sample processing data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SampleProcessing-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SampleProcessing-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
