% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_molecule_feature.R
\docType{class}
\name{SmallMoleculeFeature}
\alias{SmallMoleculeFeature}
\title{SmallMoleculeFeature}
\format{
An \code{R6Class} generator object
}
\description{
SmallMoleculeFeature Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prefix}}{character [optional]}

\item{\code{header_prefix}}{character [optional]}

\item{\code{smf_id}}{integer}

\item{\code{sme_id_refs}}{list( integer ) [optional]}

\item{\code{sme_id_ref_ambiguity_code}}{integer [optional]}

\item{\code{adduct_ion}}{character [optional]}

\item{\code{isotopomer}}{\link{Parameter} [optional]}

\item{\code{exp_mass_to_charge}}{numeric}

\item{\code{charge}}{integer}

\item{\code{retention_time_in_seconds}}{numeric [optional]}

\item{\code{retention_time_in_seconds_start}}{numeric [optional]}

\item{\code{retention_time_in_seconds_end}}{numeric [optional]}

\item{\code{abundance_assay}}{list( numeric ) [optional]}

\item{\code{opt}}{list( \link{OptColumnMapping} ) [optional]}

\item{\code{comment}}{list( \link{Comment} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SmallMoleculeFeature-new}{\code{SmallMoleculeFeature$new()}}
\item \href{#method-SmallMoleculeFeature-toJSON}{\code{SmallMoleculeFeature$toJSON()}}
\item \href{#method-SmallMoleculeFeature-fromJSON}{\code{SmallMoleculeFeature$fromJSON()}}
\item \href{#method-SmallMoleculeFeature-toJSONString}{\code{SmallMoleculeFeature$toJSONString()}}
\item \href{#method-SmallMoleculeFeature-fromJSONString}{\code{SmallMoleculeFeature$fromJSONString()}}
\item \href{#method-SmallMoleculeFeature-toDataFrame}{\code{SmallMoleculeFeature$toDataFrame()}}
\item \href{#method-SmallMoleculeFeature-fromDataFrame}{\code{SmallMoleculeFeature$fromDataFrame()}}
\item \href{#method-SmallMoleculeFeature-clone}{\code{SmallMoleculeFeature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeFeature-new"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeFeature-new}{}}}
\subsection{Method \code{new()}}{
Create a new SmallMoleculeFeature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$new(
  smf_id,
  exp_mass_to_charge,
  charge,
  prefix = "SMF",
  header_prefix = "SFH",
  sme_id_refs = NULL,
  sme_id_ref_ambiguity_code = NULL,
  adduct_ion = NULL,
  isotopomer = NULL,
  retention_time_in_seconds = NULL,
  retention_time_in_seconds_start = NULL,
  retention_time_in_seconds_end = NULL,
  abundance_assay = NULL,
  opt = NULL,
  comment = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{smf_id}}{The small molecule feature id.}

\item{\code{exp_mass_to_charge}}{The experimental mass to charge of the feature.}

\item{\code{charge}}{The charge of the feature.}

\item{\code{prefix}}{'SMF'.}

\item{\code{header_prefix}}{'SFH'.}

\item{\code{sme_id_refs}}{References by id to \link{SmallMoleculeEvidence}s.}

\item{\code{sme_id_ref_ambiguity_code}}{Ambiguity code for SME id refs.}

\item{\code{adduct_ion}}{The adduct ion.}

\item{\code{isotopomer}}{The isotopomer.}

\item{\code{retention_time_in_seconds}}{The feature's retention time in seconds.}

\item{\code{retention_time_in_seconds_start}}{The feature's retention time start in seconds.}

\item{\code{retention_time_in_seconds_end}}{The feature's retention time end in seconds.}

\item{\code{abundance_assay}}{The abundances over all assays.}

\item{\code{opt}}{Optional columns and values.}

\item{\code{comment}}{Comments.}

\item{\code{...}}{local optional variable arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeFeature-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeFeature-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeFeature-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeFeature-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$fromJSON(SmallMoleculeFeatureJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SmallMoleculeFeatureJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeFeature-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeFeature-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeFeature-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeFeature-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$fromJSONString(SmallMoleculeFeatureJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SmallMoleculeFeatureJson}}{SmallMoleculeFeatureJson string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeFeature-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeFeature-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeFeature-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeFeature-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from feature data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$fromDataFrame(FeatureDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FeatureDataFrame}}{Feature data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeFeature-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeFeature-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
