% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_molecule_summary.R
\docType{class}
\name{SmallMoleculeSummary}
\alias{SmallMoleculeSummary}
\title{SmallMoleculeSummary}
\format{
An \code{R6Class} generator object
}
\description{
SmallMoleculeSummary Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prefix}}{character [optional]}

\item{\code{header_prefix}}{character [optional]}

\item{\code{sml_id}}{integer}

\item{\code{smf_id_refs}}{list( integer ) [optional]}

\item{\code{database_identifier}}{list( character ) [optional]}

\item{\code{chemical_formula}}{list( character ) [optional]}

\item{\code{smiles}}{list( character ) [optional]}

\item{\code{inchi}}{list( character ) [optional]}

\item{\code{chemical_name}}{list( character ) [optional]}

\item{\code{uri}}{list( character ) [optional]}

\item{\code{theoretical_neutral_mass}}{list( numeric ) [optional]}

\item{\code{adduct_ions}}{list( character ) [optional]}

\item{\code{reliability}}{character [optional]}

\item{\code{best_id_confidence_measure}}{\link{Parameter} [optional]}

\item{\code{best_id_confidence_value}}{numeric [optional]}

\item{\code{abundance_assay}}{list( numeric ) [optional]}

\item{\code{abundance_study_variable}}{list( numeric ) [optional]}

\item{\code{abundance_variation_study_variable}}{list( numeric ) [optional]}

\item{\code{opt}}{list( \link{OptColumnMapping} ) [optional]}

\item{\code{comment}}{list( \link{Comment} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SmallMoleculeSummary-new}{\code{SmallMoleculeSummary$new()}}
\item \href{#method-SmallMoleculeSummary-toJSON}{\code{SmallMoleculeSummary$toJSON()}}
\item \href{#method-SmallMoleculeSummary-fromJSON}{\code{SmallMoleculeSummary$fromJSON()}}
\item \href{#method-SmallMoleculeSummary-toJSONString}{\code{SmallMoleculeSummary$toJSONString()}}
\item \href{#method-SmallMoleculeSummary-fromJSONString}{\code{SmallMoleculeSummary$fromJSONString()}}
\item \href{#method-SmallMoleculeSummary-toDataFrame}{\code{SmallMoleculeSummary$toDataFrame()}}
\item \href{#method-SmallMoleculeSummary-fromDataFrame}{\code{SmallMoleculeSummary$fromDataFrame()}}
\item \href{#method-SmallMoleculeSummary-clone}{\code{SmallMoleculeSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeSummary-new"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeSummary-new}{}}}
\subsection{Method \code{new()}}{
Create a new SmallMoleculeSummary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$new(
  sml_id,
  prefix = "SML",
  header_prefix = "SMH",
  smf_id_refs = NULL,
  database_identifier = NULL,
  chemical_formula = NULL,
  smiles = NULL,
  inchi = NULL,
  chemical_name = NULL,
  uri = NULL,
  theoretical_neutral_mass = NULL,
  adduct_ions = NULL,
  reliability = NULL,
  best_id_confidence_measure = NULL,
  best_id_confidence_value = NULL,
  abundance_assay = NULL,
  abundance_study_variable = NULL,
  abundance_variation_study_variable = NULL,
  opt = NULL,
  comment = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sml_id}}{The small molecule summary id.}

\item{\code{prefix}}{'SML'.}

\item{\code{header_prefix}}{'SMH'.}

\item{\code{smf_id_refs}}{References by id to \link{SmallMoleculeFeature}s.}

\item{\code{database_identifier}}{The database identifiers.}

\item{\code{chemical_formula}}{The chemical formulas.}

\item{\code{smiles}}{The SMILES strings.}

\item{\code{inchi}}{The INCHI identifiers.}

\item{\code{chemical_name}}{The chemical names.}

\item{\code{uri}}{External URIs.}

\item{\code{theoretical_neutral_mass}}{The theoretical neutral masses.}

\item{\code{adduct_ions}}{The adduct ions.}

\item{\code{reliability}}{The reliability according to the system defined in the \link{Metadata} section.}

\item{\code{best_id_confidence_measure}}{The best id confidence measure \link{Parameter}.}

\item{\code{best_id_confidence_value}}{The best id confidence value.}

\item{\code{abundance_assay}}{The abundances over all assays.}

\item{\code{abundance_study_variable}}{The abundances over study variables.}

\item{\code{abundance_variation_study_variable}}{The abundances variation over study variables.}

\item{\code{opt}}{Optional columns and values.}

\item{\code{comment}}{Comments.}

\item{\code{...}}{local optional variable arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeSummary-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeSummary-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeSummary-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeSummary-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$fromJSON(SmallMoleculeSummaryJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SmallMoleculeSummaryJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeSummary-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeSummary-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeSummary-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeSummary-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$fromJSONString(SmallMoleculeSummaryJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SmallMoleculeSummaryJson}}{SmallMoleculeSummaryJson string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeSummary-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeSummary-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeSummary-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeSummary-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from summary data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$fromDataFrame(SummaryDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SummaryDataFrame}}{Summary data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmallMoleculeSummary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SmallMoleculeSummary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
