% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/software.R
\docType{class}
\name{Software}
\alias{Software}
\title{Software}
\format{
An \code{R6Class} generator object
}
\description{
Software Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{parameter}}{\link{Parameter} [optional]}

\item{\code{setting}}{list( character ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Software-new}{\code{Software$new()}}
\item \href{#method-Software-toJSON}{\code{Software$toJSON()}}
\item \href{#method-Software-fromJSON}{\code{Software$fromJSON()}}
\item \href{#method-Software-toJSONString}{\code{Software$toJSONString()}}
\item \href{#method-Software-fromJSONString}{\code{Software$fromJSONString()}}
\item \href{#method-Software-toDataFrame}{\code{Software$toDataFrame()}}
\item \href{#method-Software-fromDataFrame}{\code{Software$fromDataFrame()}}
\item \href{#method-Software-clone}{\code{Software$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Software-new"></a>}}
\if{latex}{\out{\hypertarget{method-Software-new}{}}}
\subsection{Method \code{new()}}{
Create a Software
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Software$new(id = NULL, parameter = NULL, setting = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Software id.}

\item{\code{parameter}}{CV parameter for software \link{Parameter}.}

\item{\code{setting}}{Software settings list( character ).}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Software-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Software-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Software$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Software-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Software-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Software$fromJSON(SoftwareJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SoftwareJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Software-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Software-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Software$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Software-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Software-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Software$fromJSONString(SoftwareJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SoftwareJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Software-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Software-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Software$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Software-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Software-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from software data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Software$fromDataFrame(SoftwareDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SoftwareDataFrame}}{Software data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Software-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Software-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Software$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
