% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/software_all_of.R
\docType{class}
\name{SoftwareAllOf}
\alias{SoftwareAllOf}
\title{SoftwareAllOf}
\format{An \code{R6Class} generator object}
\description{
SoftwareAllOf Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parameter}}{\link{Parameter} [optional]}

\item{\code{setting}}{list( character ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SoftwareAllOf$new()}}
\item \href{#method-toJSON}{\code{SoftwareAllOf$toJSON()}}
\item \href{#method-fromJSON}{\code{SoftwareAllOf$fromJSON()}}
\item \href{#method-toJSONString}{\code{SoftwareAllOf$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SoftwareAllOf$fromJSONString()}}
\item \href{#method-clone}{\code{SoftwareAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoftwareAllOf$new(parameter = NULL, setting = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoftwareAllOf$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoftwareAllOf$fromJSON(SoftwareAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoftwareAllOf$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoftwareAllOf$fromJSONString(SoftwareAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoftwareAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
