% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringList.r
\docType{data}
\name{StringList}
\alias{StringList}
\title{StringList Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StringList
}
\description{
A typed list of strings.
}
\keyword{datasets}
