% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_variable.R
\docType{class}
\name{StudyVariable}
\alias{StudyVariable}
\title{StudyVariable}
\format{
An \code{R6Class} generator object
}
\description{
StudyVariable Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer}

\item{\code{name}}{character}

\item{\code{assay_refs}}{list( \link{Assay} ) [optional]}

\item{\code{average_function}}{\link{Parameter} [optional]}

\item{\code{variation_function}}{\link{Parameter} [optional]}

\item{\code{description}}{character [optional]}

\item{\code{factors}}{list( \link{Parameter} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudyVariable-new}{\code{StudyVariable$new()}}
\item \href{#method-StudyVariable-toJSON}{\code{StudyVariable$toJSON()}}
\item \href{#method-StudyVariable-fromJSON}{\code{StudyVariable$fromJSON()}}
\item \href{#method-StudyVariable-toJSONString}{\code{StudyVariable$toJSONString()}}
\item \href{#method-StudyVariable-fromJSONString}{\code{StudyVariable$fromJSONString()}}
\item \href{#method-StudyVariable-toDataFrame}{\code{StudyVariable$toDataFrame()}}
\item \href{#method-StudyVariable-fromDataFrame}{\code{StudyVariable$fromDataFrame()}}
\item \href{#method-StudyVariable-clone}{\code{StudyVariable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyVariable-new"></a>}}
\if{latex}{\out{\hypertarget{method-StudyVariable-new}{}}}
\subsection{Method \code{new()}}{
Create a StudyVariable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$new(
  id,
  name,
  assay_refs = NULL,
  average_function = NULL,
  variation_function = NULL,
  description = NULL,
  factors = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{StudyVariable id.}

\item{\code{name}}{StudyVariable name.}

\item{\code{assay_refs}}{Reference to list( \link{Assay} ).}

\item{\code{average_function}}{The averaging function used to calculate abundance_study_variable, defined by \link{Parameter}.}

\item{\code{variation_function}}{The variation function used to calculate abundance_variation_study_variable, defined by \link{Parameter}.}

\item{\code{description}}{Description of the StudyVariable.}

\item{\code{factors}}{A list of study variable factors list( \link{Parameter} ).}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyVariable-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StudyVariable-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyVariable-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-StudyVariable-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$fromJSON(StudyVariableJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{StudyVariableJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyVariable-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StudyVariable-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyVariable-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-StudyVariable-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$fromJSONString(StudyVariableJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{StudyVariableJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyVariable-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-StudyVariable-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyVariable-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-StudyVariable-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from StudyVariable data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$fromDataFrame(StudyVariableDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{StudyVariableDataFrame}}{StudyVariable data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyVariable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudyVariable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
