% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri.R
\docType{class}
\name{Uri}
\alias{Uri}
\title{Uri}
\format{
An \code{R6Class} generator object
}
\description{
Uri Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{value}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Uri-new}{\code{Uri$new()}}
\item \href{#method-Uri-toJSON}{\code{Uri$toJSON()}}
\item \href{#method-Uri-fromJSON}{\code{Uri$fromJSON()}}
\item \href{#method-Uri-toJSONString}{\code{Uri$toJSONString()}}
\item \href{#method-Uri-fromJSONString}{\code{Uri$fromJSONString()}}
\item \href{#method-Uri-toDataFrame}{\code{Uri$toDataFrame()}}
\item \href{#method-Uri-fromDataFrame}{\code{Uri$fromDataFrame()}}
\item \href{#method-Uri-clone}{\code{Uri$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uri-new"></a>}}
\if{latex}{\out{\hypertarget{method-Uri-new}{}}}
\subsection{Method \code{new()}}{
Create a URI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uri$new(id = NULL, value = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{URI id.}

\item{\code{value}}{The URI.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uri-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Uri-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uri$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uri-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Uri-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uri$fromJSON(UriJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UriJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uri-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Uri-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uri$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uri-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Uri-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uri$fromJSONString(UriJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UriJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uri-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Uri-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uri$toDataFrame(type = "uri")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The prefix of the URI}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uri-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-Uri-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from assay data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uri$fromDataFrame(UriDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{UriDataFrame}}{Uri data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uri-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Uri-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uri$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
