% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_api.R
\docType{class}
\name{ValidateApi}
\alias{ValidateApi}
\title{Validate operations}
\format{
An \code{R6Class} generator object
}
\description{
rmzTabM.Validate
}
\section{Methods}{

\describe{
\strong{ ValidateMzTabFile } \emph{  }
Validates an mzTab file in XML or JSON representation and reports syntactic, structural, and semantic errors.

\itemize{
\item \emph{ @param } mztabfile \link{MzTab}
\item \emph{ @param } level Enum < [info, warn, error] >
\item \emph{ @param } max.errors integer
\item \emph{ @param } semantic.validation character
\item \emph{ @returnType } list( \link{ValidationMessage} ) \cr


\item status code : 200 | Validation Okay

\item return type : array[ValidationMessage]
\item response headers :

\tabular{ll}{
}
\item status code : 415 | Unsupported content type


\item response headers :

\tabular{ll}{
}
\item status code : 422 | Invalid input

\item return type : array[ValidationMessage]
\item response headers :

\tabular{ll}{
}
\item status code : 0 | Unexpected error

\item return type : Error
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  ValidateMzTabFile  ####################

library(rmzTabM)
# MzTab | mzTab file that should be validated.
var.mztabfile <- MzTab$new()
# character | The level of errors that should be reported, one of ERROR, WARN, INFO.
var.level <- 'info' 
# integer | The maximum number of errors to return.
var.max.errors <- 100 
# character | Whether a semantic validation against the default rule set should be performed.
var.semantic.validation <- FALSE 

api.instance <- ValidateApi$new()

result <- api.instance$ValidateMzTabFile(var.mztabfile, 
  level=var.level, max.errors=var.max.errors, 
  semantic.validation=var.semantic.validation
)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ValidateApi-new}{\code{ValidateApi$new()}}
\item \href{#method-ValidateApi-ValidateMzTabFile}{\code{ValidateApi$ValidateMzTabFile()}}
\item \href{#method-ValidateApi-ValidateMzTabFileWithHttpInfo}{\code{ValidateApi$ValidateMzTabFileWithHttpInfo()}}
\item \href{#method-ValidateApi-clone}{\code{ValidateApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidateApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-ValidateApi-new}{}}}
\subsection{Method \code{new()}}{
Create a new ValidateApi object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidateApi$new(apiClient)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{apiClient}}{An initialized \link{ApiClient}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidateApi-ValidateMzTabFile"></a>}}
\if{latex}{\out{\hypertarget{method-ValidateApi-ValidateMzTabFile}{}}}
\subsection{Method \code{ValidateMzTabFile()}}{
Validate an MzTab object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidateApi$ValidateMzTabFile(
  mztabfile,
  level = "info",
  max.errors = 100,
  semantic.validation = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mztabfile}}{The \link{MzTab} object to validate.}

\item{\code{level}}{The validation level.}

\item{\code{max.errors}}{The maximum number of errors.}

\item{\code{semantic.validation}}{Whether semantic validation should also be performed.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidateApi-ValidateMzTabFileWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ValidateApi-ValidateMzTabFileWithHttpInfo}{}}}
\subsection{Method \code{ValidateMzTabFileWithHttpInfo()}}{
Validate an MzTab file but include HTTP response info.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidateApi$ValidateMzTabFileWithHttpInfo(
  mztabfile,
  level = "info",
  max.errors = 100,
  semantic.validation = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mztabfile}}{The \link{MzTab} object to validate.}

\item{\code{level}}{The validation level.}

\item{\code{max.errors}}{The maximum number of errors.}

\item{\code{semantic.validation}}{Whether semantic validation should also be performed.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidateApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ValidateApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidateApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
