% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mz_tab.R
\name{validateMzTab}
\alias{validateMzTab}
\title{Validate the provided mztab object against the online mzTab-M validator.}
\usage{
validateMzTab(
  mztab,
  validationMode = "json",
  validationLevel = "info",
  maxErrors = 100,
  semanticValidation = TRUE
)
}
\arguments{
\item{mztab}{the R6 mzTab object to write.}

\item{validationMode}{if "json", validate against the mzTab-M validator online service by converting the mzTab data model to JSON,
if "plain", validate against the mzTab-M validator online service by converting the mzTab data model to TSV format before sending.}

\item{validationLevel}{level of validation messages to return, can be one of 'info', 'warn' or 'error'.}

\item{maxErrors}{maximum number of validation errors at which the validation should stop.}

\item{semanticValidation}{use semantic validation (CV parameters) against the default mapping file.}
}
\value{
a list of validation messages or an empty list if no messages were generated.
}
\description{
Validate the provided mztab object against the online mzTab-M validator.
}
