% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_plain_api.R
\docType{class}
\name{ValidatePlainApi}
\alias{ValidatePlainApi}
\title{ValidatePlain operations}
\format{
An \code{R6Class} generator object
}
\description{
rmzTabM.ValidatePlain
}
\section{Methods}{

\describe{
\strong{ ValidatePlainMzTabFile } \emph{  }
Validates an mzTab file in plain text representation / tab-separated format and reports syntactic, structural, and semantic errors.

\itemize{
\item \emph{ @param } mztabfile character
\item \emph{ @param } level Enum < [info, warn, error] >
\item \emph{ @param } max.errors integer
\item \emph{ @param } semantic.validation character
\item \emph{ @returnType } list( \link{ValidationMessage} ) \cr


\item status code : 200 | Validation Okay

\item return type : array[ValidationMessage]
\item response headers :

\tabular{ll}{
}
\item status code : 415 | Unsupported content type


\item response headers :

\tabular{ll}{
}
\item status code : 422 | Invalid input

\item return type : array[ValidationMessage]
\item response headers :

\tabular{ll}{
}
\item status code : 0 | Unexpected error

\item return type : Error
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  ValidatePlainMzTabFile  ####################

library(rmzTabM)
# character | mzTab string content that should be validated.
var.mztabfile <- 'mztabfile_example' 
# character | The level of errors that should be reported, one of ERROR, WARN, INFO.
var.level <- 'info' 
# integer | The maximum number of errors to return.
var.max.errors <- 100 
# character | Whether a semantic validation against the default rule set should be performed.
var.semantic.validation <- FALSE 

api.instance <- ValidatePlainApi$new()

result <- api.instance$ValidatePlainMzTabFile(var.mztabfile, level=var.level,
 max.errors=var.max.errors, semantic.validation=var.semantic.validation)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ValidatePlainApi-new}{\code{ValidatePlainApi$new()}}
\item \href{#method-ValidatePlainApi-ValidatePlainMzTabFile}{\code{ValidatePlainApi$ValidatePlainMzTabFile()}}
\item \href{#method-ValidatePlainApi-ValidatePlainMzTabFileWithHttpInfo}{\code{ValidatePlainApi$ValidatePlainMzTabFileWithHttpInfo()}}
\item \href{#method-ValidatePlainApi-clone}{\code{ValidatePlainApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidatePlainApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-ValidatePlainApi-new}{}}}
\subsection{Method \code{new()}}{
Create a new ValidatePlainApi object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidatePlainApi$new(apiClient)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{apiClient}}{An initialized \link{ApiClient}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidatePlainApi-ValidatePlainMzTabFile"></a>}}
\if{latex}{\out{\hypertarget{method-ValidatePlainApi-ValidatePlainMzTabFile}{}}}
\subsection{Method \code{ValidatePlainMzTabFile()}}{
Validate a plaintext MzTab file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidatePlainApi$ValidatePlainMzTabFile(
  mztabfile,
  level = "info",
  max.errors = 100,
  semantic.validation = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mztabfile}}{The mzTab file string to validate.}

\item{\code{level}}{The validation level.}

\item{\code{max.errors}}{The maximum number of errors.}

\item{\code{semantic.validation}}{Whether semantic validation should also be performed.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidatePlainApi-ValidatePlainMzTabFileWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ValidatePlainApi-ValidatePlainMzTabFileWithHttpInfo}{}}}
\subsection{Method \code{ValidatePlainMzTabFileWithHttpInfo()}}{
Validate an plaintext MzTab file but include HTTP response info.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidatePlainApi$ValidatePlainMzTabFileWithHttpInfo(
  mztabfile,
  level = "info",
  max.errors = 100,
  semantic.validation = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mztabfile}}{The mzTab file string to validate.}

\item{\code{level}}{The validation level.}

\item{\code{max.errors}}{The maximum number of errors.}

\item{\code{semantic.validation}}{Whether semantic validation should also be performed.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidatePlainApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ValidatePlainApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidatePlainApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
