% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_tab_utils.R
\name{valueOrDefault}
\alias{valueOrDefault}
\title{Check if x is null. If so, return defaultValue, otherwise run function on x
and return the result. If a value x is encountered that is NULL, and nullable is set to FALSE,
execution will be stopped and an error message will be printed.}
\usage{
valueOrDefault(x, FUN = identity, nullable = TRUE, defaultValue = "null", ...)
}
\arguments{
\item{x}{the argument to create a value for.}

\item{FUN}{the function to apply to x, if x is not NULL.}

\item{nullable}{if TRUE, NULL is a valid value, otherwise, an error will be raised.}

\item{defaultValue}{the value to return, if x is NULL.}

\item{...}{arguments that will be passed on to FUN.}
}
\description{
Check if x is null. If so, return defaultValue, otherwise run function on x
and return the result. If a value x is encountered that is NULL, and nullable is set to FALSE,
execution will be stopped and an error message will be printed.
}
