% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mz_tab.R
\name{writeMzTab}
\alias{writeMzTab}
\title{Write MzTab object as TSV to file.}
\usage{
writeMzTab(mztab, file)
}
\arguments{
\item{mztab}{the R6 \link{MzTab} object to write.}

\item{file}{the file path to write to.}
}
\description{
Write an mzTab-M tab separated file from the passed in \link{MzTab} object.
}
