% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mz_tab.R
\name{writeMzTabJSON}
\alias{writeMzTabJSON}
\title{Converts the provided mztab object to JSON and writes it to the provided file.}
\usage{
writeMzTabJSON(mztab, file)
}
\arguments{
\item{mztab}{the R6 mzTab object to write.}

\item{file}{the file path to write to.}
}
\description{
Converts the provided mztab object to JSON and writes it to the provided file.
}
