% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzTabWriter.r
\name{writeMzTabJSONToFile}
\alias{writeMzTabJSONToFile}
\title{Converts the provided mztab object to JSON and writes it to the provided file.}
\usage{
writeMzTabJSONToFile(mztab, file)
}
\arguments{
\item{mztab}{the R6 mzTab object to write.}

\item{file}{the file path to write to.}
}
\description{
Converts the provided mztab object to JSON and writes it to the provided file.
}
