% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assay.r
\docType{data}
\name{Assay}
\alias{Assay}
\title{Assay Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Assay
}
\description{
Specification of assay. (empty) name: A name for each assay, to serve as a list of the assays that MUST be reported in the following tables.  custom: Additional custom parameters or values for a given assay.  external_uri: An external reference uri to further information about the assay, for example via a reference to an object within an ISA-TAB file.  sample_ref: An association from a given assay to the sample analysed.  ms_run_ref: An association from a given assay to the source MS run. All assays MUST reference exactly one ms_run unless a workflow with pre-fractionation is being encoded, in which case each assay MUST reference n ms_runs where n fractions have been collected. Multiple assays SHOULD reference the same ms_run to capture multiplexed experimental designs.
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{name}}{}

\item{\code{custom}}{}

\item{\code{external_uri}}{}

\item{\code{sample_ref}}{}

\item{\code{ms_run_ref}}{}
}}

\keyword{datasets}
