% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColumnParameterMapping.r
\docType{data}
\name{ColumnParameterMapping}
\alias{ColumnParameterMapping}
\title{ColumnParameterMapping Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ColumnParameterMapping
}
\description{
Defines the used unit for a column in the mzTab-M file. The format of the value has to be \\{column name}&#x3D;\\{Parameter defining the unit}. This field MUST NOT be used to define a unit for quantification columns. The unit used for small molecule quantification values MUST be set in small_molecule-quantification_unit.
}
\section{Fields}{

\describe{
\item{\code{column_name}}{}

\item{\code{param}}{mzTab-M specification example(s)
\preformatted{
COM    colunit for optional small molecule summary column with the name 'opt_global_cv_MS:MS:1002954_collisional_cross_sectional_area'
MTD    colunit-small_molecule    opt_global_cv_MS:MS:1002954_collisional_cross_sectional_area=[UO,UO:00003241, square angstrom,]

}}
}}

\keyword{datasets}
