% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_parameter_mapping.R
\docType{class}
\name{ColumnParameterMapping}
\alias{ColumnParameterMapping}
\title{ColumnParameterMapping}
\format{An \code{R6Class} generator object}
\description{
ColumnParameterMapping Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{column_name}}{character}

\item{\code{param}}{\link{Parameter}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ColumnParameterMapping$new()}}
\item \href{#method-toJSON}{\code{ColumnParameterMapping$toJSON()}}
\item \href{#method-fromJSON}{\code{ColumnParameterMapping$fromJSON()}}
\item \href{#method-toJSONString}{\code{ColumnParameterMapping$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ColumnParameterMapping$fromJSONString()}}
\item \href{#method-clone}{\code{ColumnParameterMapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnParameterMapping$new(column_name, param, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnParameterMapping$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnParameterMapping$fromJSON(ColumnParameterMappingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnParameterMapping$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnParameterMapping$fromJSONString(ColumnParameterMappingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnParameterMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
