% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comment.r
\docType{data}
\name{Comment}
\alias{Comment}
\title{Comment Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Comment
}
\description{
Comment lines can be placed anywhere in an mzTab file. These lines must start with the three-letter code COM and are ignored by most parsers. Empty lines can also occur anywhere in an mzTab file and are ignored.
}
\section{Fields}{

\describe{
\item{\code{prefix}}{}

\item{\code{msg}}{}

\item{\code{line_number}}{mzTab-M specification example(s)
\preformatted{
COM    This is a comment line

}}
}}

\keyword{datasets}
