% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contact.r
\docType{data}
\name{Contact}
\alias{Contact}
\title{Contact Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Contact
}
\description{
The contact’s name, affiliation and e-mail. Several contacts can be given by indicating the number in the square brackets after \&quot;contact\&quot;. A contact has to be supplied in the format [first name] [initials] [last name].
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{name}}{}

\item{\code{affiliation}}{}

\item{\code{email}}{mzTab-M specification example(s)
\preformatted{
MTD    contact[1]-name    James D. Watson
MTD    contact[1]-affiliation    Cambridge University, UK
MTD    contact[1]-email    watson@cam.ac.uk
MTD    contact[2]-name    Francis Crick
MTD    contact[2]-affiliation    Cambridge University, UK
MTD    contact[2]-email    crick@cam.ac.uk

}}
}}

\keyword{datasets}
