% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMzTab2MAF.R
\name{convertMzTab2MAF}
\alias{convertMzTab2MAF}
\title{convertMzTab2MAF}
\usage{
convertMzTab2MAF(mzTabfile, MAFfile)
}
\arguments{
\item{mzTabfile}{path to an mzTab-M 2.0 file to be read with rmzTab-M}

\item{MAFfile}{for an output}
}
\description{
Convert from mzTab-M to MetaboLights MAF
}
\examples{
\dontrun{
example from the vignette
}
}
