% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.r
\docType{data}
\name{CV}
\alias{CV}
\title{CV Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CV
}
\description{
Specification of controlled vocabularies. label: A string describing the labels of the controlled vocabularies/ontologies used in the mzTab file as a short-hand e.g. \&quot;MS\&quot; for PSI-MS. full_name: A string describing the full names of the controlled vocabularies/ontologies used in the mzTab file. version: A string describing the version of the controlled vocabularies/ontologies used in the mzTab file. uri: A string containing the URIs of the controlled vocabularies/ontologies used in the mzTab file.
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{label}}{}

\item{\code{full_name}}{}

\item{\code{version}}{}

\item{\code{uri}}{mzTab-M specification example(s)
\preformatted{
MTD    cv[1]-label    MS
MTD    cv[1]-full_name    PSI-MS controlled vocabulary
MTD    cv[1]-version    4.1.11
MTD    cv[1]-uri    https://raw.githubusercontent.com/HUPO-PSI/psi-ms-CV/master/psi-ms.obo
}}
}}

\keyword{datasets}
