% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.r
\docType{data}
\name{Database}
\alias{Database}
\title{Database Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Database
}
\description{
Specification of databases. (empty): The description of databases used. For cases, where a known database has not been used for identification, a userParam SHOULD be inserted to describe any identification performed e.g. de novo. If no identification has been performed at all then \&quot;no database\&quot; should be inserted followed by null. prefix: The prefix used in the “identifier” column of data tables. For the “no database” case \&quot;null\&quot; must be used. version: The database version is mandatory where identification has been performed. This may be a formal version number e.g. “1.4.1”, a date of access “2016-10-27” (ISO-8601 format) or “Unknown” if there is no suitable version that can be annotated. uri: The URI to the database. For the “no database” case, \&quot;null\&quot; must be reported.
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{param}}{}

\item{\code{prefix}}{}

\item{\code{version}}{}

\item{\code{uri}}{mzTab-M specification example(s)
\preformatted{
MTD    database[1]    [MIRIAM, MIR:00100079, HMDB, ]
MTD    database[1]-prefix    hmdb
MTD    database[1]-version    3.6
MTD    database[1]-uri    http://www.hmdb.ca/
MTD    database[2]    [,, "de novo", ]
MTD    database[2]-prefix    dn
MTD    database[2]-version    Unknown
MTD    database[2]-uri    null
MTD    database[3]    [,, "no database", null ]
MTD    database[3]-prefix    null
MTD    database[3]-version    Unknown
MTD    database[3]-uri    null

}}
}}

\keyword{datasets}
