% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mz_tab.R
\name{extractMetadata}
\alias{extractMetadata}
\title{Read the metadata from an mztab data frame and return the Metadata object.}
\usage{
extractMetadata(MzTabDataFrame)
}
\arguments{
\item{the}{mztab data frame}
}
\description{
Read the metadata from an mztab data frame and return the Metadata object.
}
