% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mz_tab.R
\name{extractMetadata}
\alias{extractMetadata}
\title{Extract metadata from data frame.}
\usage{
extractMetadata(MzTabDataFrame)
}
\arguments{
\item{MzTabDataFrame}{the mztab data frame}
}
\description{
Extract the \link{Metadata} from an mztab data frame.
}
