% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedElement.r
\docType{data}
\name{IndexedElement}
\alias{IndexedElement}
\title{IndexedElement Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IndexedElement
}
\description{
Indexed elements (IDs) define a unique ID for a collection of multiple metadata elements of the same type within the mzTab-M document, e.g. for sample, assay, study variable etc.
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{mzTab-M specification example(s)
\preformatted{
MTD    sample[1]-species[1]    [NCBITaxon, NCBITaxon:9606, Homo sapiens, ]
MTD    assay[1]    first assay description
MTD    study_variable[1]    Group B (spike-in 0.74 fmol/uL)

}}
}}

\keyword{datasets}
