% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instrument.r
\docType{data}
\name{Instrument}
\alias{Instrument}
\title{Instrument Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Instrument
}
\description{
The name, source, analyzer and detector of the instruments used in the experiment. Multiple instruments are numbered [1-n].
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{name}}{}

\item{\code{source}}{}

\item{\code{analyzer}}{}

\item{\code{detector}}{mzTab-M specification example(s)
\preformatted{
MTD    instrument[1]-name    [MS, MS:1000449, LTQ Orbitrap,]
MTD    instrument[1]-source    [MS, MS:1000073, ESI,]
…
MTD    instrument[2]-source    [MS, MS:1000598, ETD,]
MTD    instrument[1]-analyzer[1]    [MS, MS:1000291, linear ion trap,]
…
MTD    instrument[2]-analyzer[1]    [MS, MS:1000484, orbitrap,]
MTD    instrument[1]-detector    [MS, MS:1000253, electron multiplier,]
…
MTD    instrument[2]-detector    [MS, MS:1000348, focal plane collector,]

}}
}}

\keyword{datasets}
