% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.r
\docType{data}
\name{Metadata}
\alias{Metadata}
\title{Metadata Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Metadata
}
\description{
The metadata section provides additional information about the dataset(s) reported in the mzTab file. All fields in the metadata section are optional apart from those noted as mandatory. The fields in the metadata section MUST be reported in order of the various fields listed here. The field’s name and value MUST be separated by a tab character.
}
\section{Fields}{

\describe{
\item{\code{prefix}}{}

\item{\code{mzTab-version}}{}

\item{\code{mzTab-ID}}{}

\item{\code{title}}{}

\item{\code{description}}{}

\item{\code{sample_processing}}{}

\item{\code{instrument}}{}

\item{\code{software}}{}

\item{\code{publication}}{}

\item{\code{contact}}{}

\item{\code{uri}}{}

\item{\code{external_study_uri}}{}

\item{\code{quantification_method}}{}

\item{\code{study_variable}}{}

\item{\code{ms_run}}{}

\item{\code{assay}}{}

\item{\code{sample}}{}

\item{\code{custom}}{}

\item{\code{cv}}{}

\item{\code{database}}{}

\item{\code{derivatization_agent}}{}

\item{\code{small_molecule-quantification_unit}}{}

\item{\code{small_molecule_feature-quantification_unit}}{}

\item{\code{small_molecule-identification_reliability}}{}

\item{\code{id_confidence_measure}}{}

\item{\code{colunit-small_molecule}}{}

\item{\code{colunit-small_molecule_feature}}{}

\item{\code{colunit-small_molecule_evidence}}{mzTab-M specification example(s)
\preformatted{
MTD    mzTab-version    2.0.0-M
MTD    mzTab-ID    MTBL1234
MTD    title    Effects of Rapamycin on metabolite profile
...

}}
}}

\keyword{datasets}
