% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{class}
\name{Metadata}
\alias{Metadata}
\title{Metadata}
\format{
An \code{R6Class} generator object
}
\description{
Metadata Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prefix}}{character}

\item{\code{mzTab-version}}{character}

\item{\code{mzTab-ID}}{character}

\item{\code{title}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{sample_processing}}{list( \link{SampleProcessing} ) [optional]}

\item{\code{instrument}}{list( \link{Instrument} ) [optional]}

\item{\code{software}}{list( \link{Software} )}

\item{\code{publication}}{list( \link{Publication} ) [optional]}

\item{\code{contact}}{list( \link{Contact} ) [optional]}

\item{\code{uri}}{list( \link{Uri} ) [optional]}

\item{\code{external_study_uri}}{list( \link{Uri} ) [optional]}

\item{\code{quantification_method}}{\link{Parameter}}

\item{\code{study_variable}}{list( \link{StudyVariable} )}

\item{\code{ms_run}}{list( \link{MsRun} )}

\item{\code{assay}}{list( \link{Assay} )}

\item{\code{sample}}{list( \link{Sample} ) [optional]}

\item{\code{custom}}{list( \link{Parameter} ) [optional]}

\item{\code{cv}}{list( \link{CV} )}

\item{\code{database}}{list( \link{Database} )}

\item{\code{derivatization_agent}}{list( \link{Parameter} ) [optional]}

\item{\code{small_molecule-quantification_unit}}{\link{Parameter}}

\item{\code{small_molecule_feature-quantification_unit}}{\link{Parameter}}

\item{\code{small_molecule-identification_reliability}}{\link{Parameter} [optional]}

\item{\code{id_confidence_measure}}{list( \link{Parameter} )}

\item{\code{colunit-small_molecule}}{list( \link{ColumnParameterMapping} ) [optional]}

\item{\code{colunit-small_molecule_feature}}{list( \link{ColumnParameterMapping} ) [optional]}

\item{\code{colunit-small_molecule_evidence}}{list( \link{ColumnParameterMapping} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Metadata$new()}}
\item \href{#method-toJSON}{\code{Metadata$toJSON()}}
\item \href{#method-fromJSON}{\code{Metadata$fromJSON()}}
\item \href{#method-toJSONString}{\code{Metadata$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Metadata$fromJSONString()}}
\item \href{#method-toDataFrame}{\code{Metadata$toDataFrame()}}
\item \href{#method-fromDataFrame}{\code{Metadata$fromDataFrame()}}
\item \href{#method-clone}{\code{Metadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$new(
  prefix,
  `mzTab-version`,
  `mzTab-ID`,
  software,
  quantification_method,
  study_variable,
  ms_run,
  assay,
  cv,
  database,
  `small_molecule-quantification_unit`,
  `small_molecule_feature-quantification_unit`,
  id_confidence_measure,
  title = NULL,
  description = NULL,
  sample_processing = NULL,
  instrument = NULL,
  publication = NULL,
  contact = NULL,
  uri = NULL,
  external_study_uri = NULL,
  sample = NULL,
  custom = NULL,
  derivatization_agent = NULL,
  `small_molecule-identification_reliability` = NULL,
  `colunit-small_molecule` = NULL,
  `colunit-small_molecule_feature` = NULL,
  `colunit-small_molecule_evidence` = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$fromJSON(MetadataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$fromJSONString(MetadataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$fromDataFrame(MetadataDataFrame)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
