% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_run.R
\docType{class}
\name{MsRun}
\alias{MsRun}
\title{MsRun}
\format{
An \code{R6Class} generator object
}
\description{
MsRun Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer}

\item{\code{name}}{character}

\item{\code{location}}{character}

\item{\code{instrument_ref}}{\link{Instrument} [optional]}

\item{\code{format}}{\link{Parameter} [optional]}

\item{\code{id_format}}{\link{Parameter} [optional]}

\item{\code{fragmentation_method}}{list( \link{Parameter} ) [optional]}

\item{\code{scan_polarity}}{list( \link{Parameter} ) [optional]}

\item{\code{hash}}{character [optional]}

\item{\code{hash_method}}{\link{Parameter} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MsRun$new()}}
\item \href{#method-toJSON}{\code{MsRun$toJSON()}}
\item \href{#method-fromJSON}{\code{MsRun$fromJSON()}}
\item \href{#method-toJSONString}{\code{MsRun$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MsRun$fromJSONString()}}
\item \href{#method-toDataFrame}{\code{MsRun$toDataFrame()}}
\item \href{#method-fromDataFrame}{\code{MsRun$fromDataFrame()}}
\item \href{#method-clone}{\code{MsRun$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an MsRun
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRun$new(
  id,
  name,
  location,
  instrument_ref = NULL,
  format = NULL,
  id_format = NULL,
  fragmentation_method = NULL,
  scan_polarity = NULL,
  hash = NULL,
  hash_method = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{MsRun id.}

\item{\code{name}}{MsRun name.}

\item{\code{location}}{the location URI to retrieve the MsRun file.}

\item{\code{instrument_ref}}{Link to the \link{Instrument} used to acquire this MsRun.}

\item{\code{format}}{CV parameter defining the file format \link{Parameter}.}

\item{\code{id_format}}{CV parameter defining the mass spec id format \link{Parameter}.}

\item{\code{fragmentation_method}}{Fragementation methods list( \link{Parameter} ).}

\item{\code{scan_polarity}}{Scan polarities present in the file list( \link{Parameter} ).}

\item{\code{hash}}{Hash value of the file using the hash_method.}

\item{\code{hash_method}}{\link{Parameter}}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRun$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRun$fromJSON(MsRunJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MsRunJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRun$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRun$fromJSONString(MsRunJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MsRunJson}}{JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRun$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from ms run data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRun$fromDataFrame(MsRunDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MsRunDataFrame}}{MsRun data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MsRun$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
