% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTD.R
\name{mtd_skeleton}
\alias{mtd_skeleton}
\title{Create a skeleton MTD section}
\usage{
mtd_skeleton(
  id = character(),
  software = character(),
  quantification_method = "[MS, MS:1001834, LC-MS label-free quantitation analysis, ]",
  cv_label = c("MS", "PRIDE"),
  cv_full_name = c("PSI-MS controlled vocabulary",
    "PRIDE PRoteomics IDEntifications (PRIDE) database controlled vocabulary"),
  cv_version = c("4.1.138", "16:10:2023 11:38"),
  cv_uri = c("https://raw.githubusercontent.com/HUPO-PSI/psi-ms-CV/master/psi-ms.obo",
    "https://www.ebi.ac.uk/ols/ontologies/pride"),
  database = c("[,, \\"no database\\", null ]"),
  database_prefix = c("null"),
  database_version = c("Unknown"),
  database_uri = c("null"),
  small_molecule_quantification_unit =
    "[PRIDE, PRIDE:0000330, Arbitrary quantification unit, ]",
  small_molecule_feature_quantification_unit =
    "[PRIDE, PRIDE:0000330, Arbitrary quantification unit, ]",
  small_molecule_identification_reliability =
    "[MS, MS:1002896, compound identification confidence level, ]",
  mztab_version = "2.0.0-M"
)
}
\arguments{
\item{id}{`character(1)` (**mandatory**) with the ID of the data set.}

\item{software}{`character` (**mandatory**) with the software(s) used.
Can be of length > 1 if multiple softwares were used.}

\item{quantification_method}{`character(1)` defining the quantification
method used in the experiment.}

\item{cv_label}{`character` describing the labels of the controlled
vocabularies/ontologies used in the mzTab file as a short-hand, e.g.
`cv_label = "MS"` for PSI-MS.}

\item{cv_full_name}{`character` with the full names of the controlled
vocabularies/ontologies used in the mzTab file.}

\item{cv_version}{`character` with the version of the used
vocabularies/ontologies.}

\item{cv_uri}{`character` with the URIs of the vocabularies/ontologies.}

\item{database}{`character` defining the database used for annotation. If no
annotation/identification was performed then `"[,, no database, null]"`
should be used.}

\item{database_prefix}{`character` defining the prefix used in the
*identifier* column of data tables. For *no database*, `"null"` must
be used.}

\item{database_version}{`character` with the database version used.}

\item{database_uri}{`character` with the URI to the database(s). For
*no database* `"null"` must be used.}

\item{small_molecule_quantification_unit}{`character(1)` defines the type
of units are reported in the small molecule summary quantification/
abundance fields.}

\item{small_molecule_feature_quantification_unit}{`character(1)` defines
what type of units are reported in the small molecule feature
quantification / abundance fields.}

\item{small_molecule_identification_reliability}{`character(1)` defines the
system used for giving reliability / confidence codes to small molecule
identifications MUST be specified if not using the default codes.}

\item{mztab_version}{`character(1)` defining the mzTab-M version of the file.}
}
\value{
two-column `character` `matrix` that should be expanded with
    additional fields (such as *title*, *description* etc) and
    information (with the help from the `mtd_fields()` function).
}
\description{
Create a `matrix` with the basic mzTab-M *MTD* section based on the provided
data. The returned result contains only minimal information. It should be
expanded, corrected and completed with additional fields and information
(i.e., the *skeleton* returned by this function should be completed with
*flesh*).

For details and expected input for the various parameter it is **strongly
suggested** to consult the [mzTab-M](https://github.com/HUPO-PSI/mzTab-M/blob/main/specification_documents/mzTab_format_specification_2_1-M.adoc#62-metadata-section) documentation.
}
\examples{

## Define a minimal mzTab-M metadata information
mtd <- mtd_skeleton(id = "001", software = "[MS, MS:1001582, xmcs, 4.0.0]")

## Column 1 has the field names
mtd[, 1]

## Column 2 the respective values
mtd[, 2]

## Add additional fields as defined in the mzTab-M definition
mtd <- rbind(
    mtd,
    c("title", "My simple xcms preprocessed data"),
    c("description", "A simple example xcms preprocessing."))

tail(mtd)

## Add instrument information
instr <- mtd_fields(
    name = "[MS, MS:1000449, LTQ Orbitrap,]",
    source = "[MS, MS:1000073, ESI,]",
    `analyzer[1]` = "[MS, MS:1000291, linear ion trap,]",
    detector = "[MS, MS:1000253, electron multiplier,]",
    field_prefix = "instrument"
)
instr

## Add this information to the metadata
mtd <- rbind(mtd, instr)

## Define sample processing fields using the mtd_fields function
sp <- mtd_fields(
    c("[MSIO, MSIO:0000146, centrifugation,]",
      "[MSIO, MSIO:0000141, metabolite extraction,]",
      "[MSIO, MSIO:0000141, silylation,]"),
    field_prefix = "sample_processing")
sp

## Add this information to the metadata
mtd <- rbind(mtd, sp)

## Since a new ontology was used for the sample processing, we need also to
## add that to the metadata. We manually define the fields to add using
## `"cv[3]"` because there are already 2 CVs defined in the MTD skeleton.
cv2 <- rbind(
    c("cv[3]-label", "MSIO"),
    c("cv[3]-full_name", "Metabolomics Standards Initiative Ontology"),
    c("cv[3]-version", "1.0.1"),
    c("cv[3]-uri", "http://purl.obolibrary.org/obo/msio.owl")
)

## Add this information to the metadata
mtd <- rbind(mtd, cv2)

## Finally sort the metadata fields according to the expected order
mtd <- mtd_sort(mtd)
mtd
}
\author{
Philippine Louail, Johannes Rainer
}
