% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptColumnMapping.r
\docType{data}
\name{OptColumnMapping}
\alias{OptColumnMapping}
\title{OptColumnMapping Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OptColumnMapping
}
\description{
Additional columns can be added to the end of the small molecule table. These column headers MUST start with the prefix “opt_” followed by the {identifier} of the object they reference: assay, study variable, MS run or “global” (if the value relates to all replicates). Column names MUST only contain the following characters: ‘A’-‘Z’, ‘a’-‘z’, ‘0’-‘9’, ‘’, ‘-’, ‘[’, ‘]’, and ‘:’. CV parameter accessions MAY be used for optional columns following the format: opt{identifier}_cv_{accession}_\\{parameter name}. Spaces within the parameter’s name MUST be replaced by ‘_’.
}
\section{Fields}{

\describe{
\item{\code{identifier}}{}

\item{\code{param}}{}

\item{\code{value}}{mzTab-M specification example(s)
\preformatted{
SMH    SML_ID    …    opt_assay[1]_my_value    …    opt_global_another_value
SML    1    …    My value    …    some other value

}}
}}

\keyword{datasets}
