% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.r
\docType{data}
\name{Parameter}
\alias{Parameter}
\title{Parameter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Parameter
}
\description{
mzTab makes use of CV parameters. As mzTab is expected to be used in several experimental environments where parameters might not yet be available for the generated scores etc. all parameters can either report CV parameters or user parameters that only contain a name and a value. Parameters are always reported as [CV label, accession, name, value]. Any field that is not available MUST be left empty.
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{cv_label}}{}

\item{\code{cv_accession}}{}

\item{\code{name}}{}

\item{\code{value}}{mzTab-M specification example(s)
\preformatted{
[MS, MS:1001477, SpectraST,]
[,,A user parameter, The value]

}}
}}

\keyword{datasets}
