% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publication.r
\docType{data}
\name{Publication}
\alias{Publication}
\title{Publication Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Publication
}
\description{
A publication associated with this file. Several publications can be given by indicating the number in the square brackets after “publication”. PubMed ids must be prefixed by “pubmed:”, DOIs by “doi:”. Multiple identifiers MUST be separated by “|”.
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{publicationItems}}{mzTab-M specification example(s)
\preformatted{
MTD    publication[1]    pubmed:21063943|doi:10.1007/978-1-60761-987-1_6
MTD    publication[2]    pubmed:20615486|doi:10.1016/j.jprot.2010.06.008

}}
}}

\keyword{datasets}
