% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PublicationItem.r
\docType{data}
\name{PublicationItem}
\alias{PublicationItem}
\title{PublicationItem Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PublicationItem
}
\description{
A publication item, defined by a qualifier and a native accession, e.g. pubmed id.
}
\section{Fields}{

\describe{
\item{\code{type}}{}

\item{\code{accession}}{}
}}

\keyword{datasets}
