% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mz_tab.R
\name{readMzTab}
\alias{readMzTab}
\title{Read mzTab-M file into data frame.}
\usage{
readMzTab(filename)
}
\arguments{
\item{filename}{the mzTab file to parse.}
}
\description{
Read an mzTab tab separated file from the passed in file.
}
