% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{safe_unbox}
\alias{safe_unbox}
\title{Extends jsonlite's safe_unbox method.}
\usage{
safe_unbox(x, default = NULL)
}
\arguments{
\item{x}{the list, vector, or array object to unbox into a singleton}

\item{default}{the default value to return, if x is null}
}
\description{
This method adds a default value to return if its argument is null in comparison to rmzTabM::safe_unbox.
}
