% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{class}
\name{Sample}
\alias{Sample}
\title{Sample}
\format{An \code{R6Class} generator object}
\description{
Sample Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer}

\item{\code{elementType}}{character}

\item{\code{name}}{character [optional]}

\item{\code{custom}}{list( \link{Parameter} ) [optional]}

\item{\code{species}}{list( \link{Parameter} ) [optional]}

\item{\code{tissue}}{list( \link{Parameter} ) [optional]}

\item{\code{cell_type}}{list( \link{Parameter} ) [optional]}

\item{\code{disease}}{list( \link{Parameter} ) [optional]}

\item{\code{description}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Sample$new()}}
\item \href{#method-toJSON}{\code{Sample$toJSON()}}
\item \href{#method-fromJSON}{\code{Sample$fromJSON()}}
\item \href{#method-toJSONString}{\code{Sample$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Sample$fromJSONString()}}
\item \href{#method-clone}{\code{Sample$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$new(
  id,
  elementType,
  name = NULL,
  custom = NULL,
  species = NULL,
  tissue = NULL,
  cell_type = NULL,
  disease = NULL,
  description = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$fromJSON(SampleJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$fromJSONString(SampleJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
