% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.r
\docType{data}
\name{Sample}
\alias{Sample}
\title{Sample Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Sample
}
\description{
Specification of sample. (empty) name: A name for each sample to serve as a list of the samples that MUST be reported in the following tables. Samples MUST be reported if a statistical design is being captured (i.e. bio or tech replicates). If the type of replicates are not known, samples SHOULD NOT be reported.  species: The respective species of the samples analysed. For more complex cases, such as metagenomics, optional columns and userParams should be used.  tissue: The respective tissue(s) of the sample.  cell_type: The respective cell type(s) of the sample.  disease: The respective disease(s) of the sample.  description: A human readable description of the sample.  custom: Custom parameters describing the sample&#39;s additional properties. Dates MUST be provided in ISO-8601 format.
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{name}}{}

\item{\code{custom}}{}

\item{\code{species}}{}

\item{\code{tissue}}{}

\item{\code{cell_type}}{}

\item{\code{disease}}{}

\item{\code{description}}{mzTab-M specification example(s)
\preformatted{
COM    Experiment where all samples consisted of the same two species
MTD    sample[1]    individual number 1
MTD    sample[1]-species[1]    [NCBITaxon, NCBITaxon:9606, Homo sapiens, ]
MTD    sample[1]-tissue[1]    [BTO, BTO:0000759, liver, ]
MTD    sample[1]-cell_type[1]    [CL, CL:0000182, hepatocyte, ]
MTD    sample[1]-disease[1]    [DOID, DOID:684, hepatocellular carcinoma, ]
MTD    sample[1]-disease[2]    [DOID, DOID:9451, alcoholic fatty liver, ]
MTD    sample[1]-description    Hepatocellular carcinoma samples.
MTD    sample[1]-custom[1]    [,,Extraction date, 2011-12-21]
MTD    sample[1]-custom[2]    [,,Extraction reason, liver biopsy]
MTD    sample[2]    individual number 2
MTD    sample[2]-species[1]    [NCBITaxon, NCBITaxon:9606, Homo sapiens, ]
MTD    sample[2]-tissue[1]    [BTO, BTO:0000759, liver, ]
MTD    sample[2]-cell_type[1]    [CL, CL:0000182, hepatocyte, ]
MTD    sample[2]-description    Healthy control samples.

}}
}}

\keyword{datasets}
