% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{class}
\name{Sample}
\alias{Sample}
\title{Sample}
\format{
An \code{R6Class} generator object
}
\description{
Sample Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{name}}{character [optional]}

\item{\code{custom}}{list( \link{Parameter} ) [optional]}

\item{\code{species}}{list( \link{Parameter} ) [optional]}

\item{\code{tissue}}{list( \link{Parameter} ) [optional]}

\item{\code{cell_type}}{list( \link{Parameter} ) [optional]}

\item{\code{disease}}{list( \link{Parameter} ) [optional]}

\item{\code{description}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Sample$new()}}
\item \href{#method-toJSON}{\code{Sample$toJSON()}}
\item \href{#method-fromJSON}{\code{Sample$fromJSON()}}
\item \href{#method-toJSONString}{\code{Sample$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Sample$fromJSONString()}}
\item \href{#method-toDataFrame}{\code{Sample$toDataFrame()}}
\item \href{#method-fromDataFrame}{\code{Sample$fromDataFrame()}}
\item \href{#method-clone}{\code{Sample$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$new(
  id = NULL,
  name = NULL,
  custom = NULL,
  species = NULL,
  tissue = NULL,
  cell_type = NULL,
  disease = NULL,
  description = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Sample id.}

\item{\code{name}}{Sample name.}

\item{\code{custom}}{Custom parameters list( \link{Parameter} ).}

\item{\code{species}}{Species CV parameters list( \link{Parameter} ).}

\item{\code{tissue}}{Tissue CV parameters list( \link{Parameter} ).}

\item{\code{cell_type}}{Cell Type CV parameters list( \link{Parameter} ).}

\item{\code{disease}}{Disease CV parameters list( \link{Parameter} ).}

\item{\code{description}}{Description of the sample.}

\item{\code{...}}{local optional variable arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize to list object suitable for jsonlite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize from jsonlite list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$fromJSON(SampleJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SampleJson}}{list object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize from JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$fromJSONString(SampleJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SampleJson}}{SampleJson string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
Serialize to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
Deserialize from sample data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$fromDataFrame(SampleDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SampleDataFrame}}{Sample data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
