% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleProcessing.r
\docType{data}
\name{SampleProcessing}
\alias{SampleProcessing}
\title{SampleProcessing Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SampleProcessing
}
\description{
A list of parameters describing a sample processing, preparation or handling step similar to a biological or analytical methods report. The order of the sample_processing items should reflect the order these processing steps were performed in. If multiple parameters are given for a step these MUST be separated by a “|”. If derivatization was performed, it MUST be reported here as a general step, e.g. &#39;silylation&#39; and the actual derivatization agens MUST be specified in the Section 6.2.54 part.
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{sampleProcessing}}{mzTab-M specification example(s)
\preformatted{
MTD    sample_processing[1]    [MSIO, MSIO:0000107, metabolism quenching using precooled 60 percent methanol ammonium bicarbonate buffer,]
MTD    sample_processing[2]    [MSIO, MSIO:0000146, centrifugation,]
MTD    sample_processing[3]    [MSIO, MSIO:0000141, metabolite extraction,]
MTD    sample_processing[4]    [MSIO, MSIO:0000141, silylation,]

}}
}}

\keyword{datasets}
