% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_processing.R
\docType{class}
\name{SampleProcessing}
\alias{SampleProcessing}
\title{SampleProcessing}
\format{
An \code{R6Class} generator object
}
\description{
SampleProcessing Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{sampleProcessing}}{list( \link{Parameter} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SampleProcessing$new()}}
\item \href{#method-toJSON}{\code{SampleProcessing$toJSON()}}
\item \href{#method-fromJSON}{\code{SampleProcessing$fromJSON()}}
\item \href{#method-toJSONString}{\code{SampleProcessing$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SampleProcessing$fromJSONString()}}
\item \href{#method-toDataFrame}{\code{SampleProcessing$toDataFrame()}}
\item \href{#method-fromDataFrame}{\code{SampleProcessing$fromDataFrame()}}
\item \href{#method-clone}{\code{SampleProcessing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$new(id = NULL, sampleProcessing = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$fromJSON(SampleProcessingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$fromJSONString(SampleProcessingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$toDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-fromDataFrame}{}}}
\subsection{Method \code{fromDataFrame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$fromDataFrame(SampleProcessingDataFrame)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SampleProcessing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
