% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallMoleculeEvidence.r
\docType{data}
\name{SmallMoleculeEvidence}
\alias{SmallMoleculeEvidence}
\title{SmallMoleculeEvidence Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SmallMoleculeEvidence
}
\description{
The small molecule evidence section is table-based, representing evidence for identifications of small molecules/features, from database search or any other process used to give putative identifications to molecules. In a typical case, each row represents one result from a single search or intepretation of a piece of evidence e.g. a database search with a fragmentation spectrum. Multiple results from a given input data item (e.g. one fragment spectrum) SHOULD share the same value under evidence_input_id.  The small molecule evidence section MUST always come after the Small Molecule Feature Table. All table columns MUST be Tab separated. There MUST NOT be any empty cells. Missing values MUST be reported using “null”.  The order of columns MUST follow the order specified below.  All columns are MANDATORY except for “opt_” columns.
}
\section{Fields}{

\describe{
\item{\code{prefix}}{}

\item{\code{header_prefix}}{}

\item{\code{sme_id}}{}

\item{\code{evidence_input_id}}{}

\item{\code{database_identifier}}{}

\item{\code{chemical_formula}}{}

\item{\code{smiles}}{}

\item{\code{inchi}}{}

\item{\code{chemical_name}}{}

\item{\code{uri}}{}

\item{\code{derivatized_form}}{}

\item{\code{adduct_ion}}{}

\item{\code{exp_mass_to_charge}}{}

\item{\code{charge}}{}

\item{\code{theoretical_mass_to_charge}}{}

\item{\code{spectra_ref}}{}

\item{\code{identification_method}}{}

\item{\code{ms_level}}{}

\item{\code{id_confidence_measure}}{}

\item{\code{rank}}{}

\item{\code{opt}}{}

\item{\code{comment}}{}
}}

\keyword{datasets}
