% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallMoleculeFeature.r
\docType{data}
\name{SmallMoleculeFeature}
\alias{SmallMoleculeFeature}
\title{SmallMoleculeFeature Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SmallMoleculeFeature
}
\description{
The small molecule feature section is table-based, representing individual MS regions (generally considered to be the elution profile for all isotopomers formed from a single charge state of a molecule), that have been measured/quantified. However, for approaches that quantify individual isotopomers e.g. stable isotope labelling/flux studies, then each SMF row SHOULD represent a single isotopomer.  Different adducts or derivatives and different charge states of individual molecules should be reported as separate SMF rows.  The small molecule feature section MUST always come after the Small Molecule Table. All table columns MUST be Tab separated. There MUST NOT be any empty cells. Missing values MUST be reported using “null”.  The order of columns MUST follow the order specified below.  All columns are MANDATORY except for “opt_” columns.
}
\section{Fields}{

\describe{
\item{\code{prefix}}{}

\item{\code{header_prefix}}{}

\item{\code{smf_id}}{}

\item{\code{sme_id_refs}}{}

\item{\code{sme_id_ref_ambiguity_code}}{}

\item{\code{adduct_ion}}{}

\item{\code{isotopomer}}{}

\item{\code{exp_mass_to_charge}}{}

\item{\code{charge}}{}

\item{\code{retention_time_in_seconds}}{}

\item{\code{retention_time_in_seconds_start}}{}

\item{\code{retention_time_in_seconds_end}}{}

\item{\code{abundance_assay}}{}

\item{\code{opt}}{}

\item{\code{comment}}{}
}}

\keyword{datasets}
