% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_molecule_feature.R
\docType{class}
\name{SmallMoleculeFeature}
\alias{SmallMoleculeFeature}
\title{SmallMoleculeFeature}
\format{
An \code{R6Class} generator object
}
\description{
SmallMoleculeFeature Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prefix}}{character [optional]}

\item{\code{header_prefix}}{character [optional]}

\item{\code{smf_id}}{integer}

\item{\code{sme_id_refs}}{list( integer ) [optional]}

\item{\code{sme_id_ref_ambiguity_code}}{integer [optional]}

\item{\code{adduct_ion}}{character [optional]}

\item{\code{isotopomer}}{\link{Parameter} [optional]}

\item{\code{exp_mass_to_charge}}{numeric}

\item{\code{charge}}{integer}

\item{\code{retention_time_in_seconds}}{numeric [optional]}

\item{\code{retention_time_in_seconds_start}}{numeric [optional]}

\item{\code{retention_time_in_seconds_end}}{numeric [optional]}

\item{\code{abundance_assay}}{list( numeric ) [optional]}

\item{\code{opt}}{list( \link{OptColumnMapping} ) [optional]}

\item{\code{comment}}{list( \link{Comment} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SmallMoleculeFeature$new()}}
\item \href{#method-toJSON}{\code{SmallMoleculeFeature$toJSON()}}
\item \href{#method-fromJSON}{\code{SmallMoleculeFeature$fromJSON()}}
\item \href{#method-toJSONString}{\code{SmallMoleculeFeature$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SmallMoleculeFeature$fromJSONString()}}
\item \href{#method-clone}{\code{SmallMoleculeFeature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$new(
  smf_id,
  exp_mass_to_charge,
  charge,
  prefix = "SMF",
  header_prefix = "SFH",
  sme_id_refs = NULL,
  sme_id_ref_ambiguity_code = NULL,
  adduct_ion = NULL,
  isotopomer = NULL,
  retention_time_in_seconds = NULL,
  retention_time_in_seconds_start = NULL,
  retention_time_in_seconds_end = NULL,
  abundance_assay = NULL,
  opt = NULL,
  comment = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$fromJSON(SmallMoleculeFeatureJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$fromJSONString(SmallMoleculeFeatureJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeFeature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
