% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallMoleculeSummary.r
\docType{data}
\name{SmallMoleculeSummary}
\alias{SmallMoleculeSummary}
\title{SmallMoleculeSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SmallMoleculeSummary
}
\description{
The small molecule summary section is table-based, represented summarized quantitative information across assays and study variables, grouped by identification in rows. The small molecule section MUST always come after the metadata section. All table columns MUST be Tab separated. There MUST NOT be any empty cells; missing values MUST be reported using “null” for columns where Is Nullable &#x3D; “True”.  Each row of the small molecule section is intended to report one final result to be communicated in terms of a molecule that has been quantified. In many cases, this may be the molecule of biological interest, although in some cases, the final result could be a derivatized form as appropriate – although it is desirable for the database identifier(s) to reference to the biological (non-derivatized) form. In general, different adduct forms would generally be reported in the Small Molecule Feature section.  The order of columns MUST follow the order specified below.  All columns are MANDATORY except for “opt_” columns.
}
\section{Fields}{

\describe{
\item{\code{prefix}}{}

\item{\code{header_prefix}}{}

\item{\code{sml_id}}{}

\item{\code{smf_id_refs}}{}

\item{\code{database_identifier}}{}

\item{\code{chemical_formula}}{}

\item{\code{smiles}}{}

\item{\code{inchi}}{}

\item{\code{chemical_name}}{}

\item{\code{uri}}{}

\item{\code{theoretical_neutral_mass}}{}

\item{\code{adduct_ions}}{}

\item{\code{reliability}}{}

\item{\code{best_id_confidence_measure}}{}

\item{\code{best_id_confidence_value}}{}

\item{\code{abundance_assay}}{}

\item{\code{abundance_study_variable}}{}

\item{\code{abundance_variation_study_variable}}{}

\item{\code{opt}}{}

\item{\code{comment}}{}
}}

\keyword{datasets}
