% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_molecule_summary.R
\docType{class}
\name{SmallMoleculeSummary}
\alias{SmallMoleculeSummary}
\title{SmallMoleculeSummary}
\format{An \code{R6Class} generator object}
\description{
SmallMoleculeSummary Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prefix}}{character [optional]}

\item{\code{header_prefix}}{character [optional]}

\item{\code{sml_id}}{integer}

\item{\code{smf_id_refs}}{list( integer ) [optional]}

\item{\code{database_identifier}}{list( character ) [optional]}

\item{\code{chemical_formula}}{list( character ) [optional]}

\item{\code{smiles}}{list( character ) [optional]}

\item{\code{inchi}}{list( character ) [optional]}

\item{\code{chemical_name}}{list( character ) [optional]}

\item{\code{uri}}{list( character ) [optional]}

\item{\code{theoretical_neutral_mass}}{list( numeric ) [optional]}

\item{\code{adduct_ions}}{list( character ) [optional]}

\item{\code{reliability}}{character [optional]}

\item{\code{best_id_confidence_measure}}{\link{Parameter} [optional]}

\item{\code{best_id_confidence_value}}{numeric [optional]}

\item{\code{abundance_assay}}{list( numeric ) [optional]}

\item{\code{abundance_study_variable}}{list( numeric ) [optional]}

\item{\code{abundance_variation_study_variable}}{list( numeric ) [optional]}

\item{\code{opt}}{list( \link{OptColumnMapping} ) [optional]}

\item{\code{comment}}{list( \link{Comment} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SmallMoleculeSummary$new()}}
\item \href{#method-toJSON}{\code{SmallMoleculeSummary$toJSON()}}
\item \href{#method-fromJSON}{\code{SmallMoleculeSummary$fromJSON()}}
\item \href{#method-toJSONString}{\code{SmallMoleculeSummary$toJSONString()}}
\item \href{#method-fromJSONString}{\code{SmallMoleculeSummary$fromJSONString()}}
\item \href{#method-clone}{\code{SmallMoleculeSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$new(
  sml_id,
  prefix = "SML",
  header_prefix = "SMH",
  smf_id_refs = NULL,
  database_identifier = NULL,
  chemical_formula = NULL,
  smiles = NULL,
  inchi = NULL,
  chemical_name = NULL,
  uri = NULL,
  theoretical_neutral_mass = NULL,
  adduct_ions = NULL,
  reliability = NULL,
  best_id_confidence_measure = NULL,
  best_id_confidence_value = NULL,
  abundance_assay = NULL,
  abundance_study_variable = NULL,
  abundance_variation_study_variable = NULL,
  opt = NULL,
  comment = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$fromJSON(SmallMoleculeSummaryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$fromJSONString(SmallMoleculeSummaryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmallMoleculeSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
