% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Software.r
\docType{data}
\name{Software}
\alias{Software}
\title{Software Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Software
}
\description{
Software used to analyze the data and obtain the reported results. The parameter’s value SHOULD contain the software’s version. The order (numbering) should reflect the order in which the tools were used. A software setting used. This field MAY occur multiple times for a single software. The value of this field is deliberately set as a String, since there currently do not exist CV terms for every possible setting.
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{parameter}}{}

\item{\code{setting}}{mzTab-M specification example(s)
\preformatted{
MTD    software[1]    [MS, MS:1002879, Progenesis QI, 3.0]
MTD    software[1]-setting    Fragment tolerance = 0.1 Da
…
MTD    software[2]-setting    Parent tolerance = 0.5 Da

}}
}}

\keyword{datasets}
