% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraRef.r
\docType{data}
\name{SpectraRef}
\alias{SpectraRef}
\title{SpectraRef Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpectraRef
}
\description{
Reference to a spectrum in a spectrum file, for example a fragmentation spectrum has been used to support the identification. If a separate spectrum file has been used for fragmentation spectrum, this MUST be reported in the metadata section as additional ms_runs. The reference must be in the format ms_run[1-n]:{SPECTRA_REF} where SPECTRA_REF MUST follow the format defined in 5.2 (including references to chromatograms where these are used to inform identification). Multiple spectra MUST be referenced using a “|” delimited list for the (rare) cases in which search engines have combined or aggregated multiple spectra in advance of the search to make identifications.  If a fragmentation spectrum has not been used, the value should indicate the ms_run to which is identification is mapped e.g. “ms_run[1]”.
}
\section{Fields}{

\describe{
\item{\code{ms_run}}{}

\item{\code{reference}}{mzTab-M specification example(s)
\preformatted{
SEH    SME_ID    …    spectra_ref    …
SME    1            ms_run[1]:index=5    …

}}
}}

\keyword{datasets}
