% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyVariable.r
\docType{data}
\name{StudyVariable}
\alias{StudyVariable}
\title{StudyVariable Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StudyVariable
}
\description{
Specification of study_variable. (empty) name: A name for each study variable (experimental condition or factor), to serve as a list of the study variables that MUST be reported in the following tables. For software that does not capture study variables, a single study variable MUST be reported, linking to all assays. This single study variable MUST have the identifier “undefined“. assay_refs: Bar-separated references to the IDs of assays grouped in the study variable. average_function: The function used to calculate the study variable quantification value and the operation used is not arithmetic mean (default) e.g. “geometric mean”, “median”. The 1-n refers to different study variables. variation_function: The function used to calculate the study variable quantification variation value if it is reported and the operation used is not coefficient of variation (default) e.g. “standard error”. description: A textual description of the study variable. factors: Additional parameters or factors, separated by bars, that are known about study variables allowing the capture of more complex, such as nested designs.
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{name}}{}

\item{\code{assay_refs}}{}

\item{\code{average_function}}{}

\item{\code{variation_function}}{}

\item{\code{description}}{}

\item{\code{factors}}{mzTab-M specification example(s)
\preformatted{
MTD    study_variable[1]    control
MTD    study_variable[1]-assay_refs    assay[1]| assay[2]| assay[3]
MTD    study_variable-average_function    [MS, MS:1002883, median, ]
MTD    study_variable-variation_function    [MS, MS:1002885, standard error, ]
MTD    study_variable[1]-description    Group B (spike-in 0.74 fmol/uL)
MTD    study_variable[1]-factors    [,,time point, 1 minute]|[,,rapamycin dose,0.5mg]
MTD    study_variable[2]    1 minute 0.5mg rapamycin

}}
}}

\keyword{datasets}
