% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_variable.R
\docType{class}
\name{StudyVariable}
\alias{StudyVariable}
\title{StudyVariable}
\format{An \code{R6Class} generator object}
\description{
StudyVariable Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer}

\item{\code{elementType}}{character}

\item{\code{name}}{character}

\item{\code{assay_refs}}{list( \link{Assay} ) [optional]}

\item{\code{average_function}}{\link{Parameter} [optional]}

\item{\code{variation_function}}{\link{Parameter} [optional]}

\item{\code{description}}{character [optional]}

\item{\code{factors}}{list( \link{Parameter} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StudyVariable$new()}}
\item \href{#method-toJSON}{\code{StudyVariable$toJSON()}}
\item \href{#method-fromJSON}{\code{StudyVariable$fromJSON()}}
\item \href{#method-toJSONString}{\code{StudyVariable$toJSONString()}}
\item \href{#method-fromJSONString}{\code{StudyVariable$fromJSONString()}}
\item \href{#method-clone}{\code{StudyVariable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$new(
  id,
  elementType,
  name,
  assay_refs = NULL,
  average_function = NULL,
  variation_function = NULL,
  description = NULL,
  factors = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$fromJSON(StudyVariableJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$fromJSONString(StudyVariableJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
