% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uri.r
\docType{data}
\name{Uri}
\alias{Uri}
\title{Uri Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Uri
}
\description{
A URI pointing to the file’s source data (e.g., a MetaboLights records) or an external file with more details about the study design.
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{elementType}}{}

\item{\code{value}}{mzTab-M specification example(s)
\preformatted{
MTD    uri[1]    https://www.ebi.ac.uk/metabolights/MTBLS517
…
MTD    external_study_uri[1]    https://www.ebi.ac.uk/metabolights/MTBLS517/files/i_Investigation.txt

}}
}}

\keyword{datasets}
