% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateApi.r
\docType{data}
\name{ValidateApi}
\alias{ValidateApi}
\title{Validate operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidateApi
}
\description{
rmztab.Validate
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

validate_mz_tab_file  

}
}

\keyword{datasets}
