% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatePlainApi.r
\docType{data}
\name{ValidatePlainApi}
\alias{ValidatePlainApi}
\title{ValidatePlain operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidatePlainApi
}
\description{
rmztab.ValidatePlain
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

validate_plain_mz_tab_file  

}
}

\keyword{datasets}
