% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_message.R
\docType{class}
\name{ValidationMessage}
\alias{ValidationMessage}
\title{ValidationMessage}
\format{An \code{R6Class} generator object}
\description{
ValidationMessage Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{code}}{character}

\item{\code{category}}{character}

\item{\code{message_type}}{character [optional]}

\item{\code{message}}{character}

\item{\code{line_number}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ValidationMessage$new()}}
\item \href{#method-toJSON}{\code{ValidationMessage$toJSON()}}
\item \href{#method-fromJSON}{\code{ValidationMessage$fromJSON()}}
\item \href{#method-toJSONString}{\code{ValidationMessage$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ValidationMessage$fromJSONString()}}
\item \href{#method-clone}{\code{ValidationMessage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationMessage$new(
  code,
  category,
  message,
  message_type = "info",
  line_number = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationMessage$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationMessage$fromJSON(ValidationMessageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationMessage$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationMessage$fromJSONString(ValidationMessageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidationMessage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
