% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationMessage.r
\docType{data}
\name{ValidationMessage}
\alias{ValidationMessage}
\title{ValidationMessage Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ValidationMessage
}
\description{
ValidationMessage Class
}
\section{Fields}{

\describe{
\item{\code{code}}{}

\item{\code{category}}{}

\item{\code{message_type}}{}

\item{\code{message}}{}

\item{\code{line_number}}{}
}}

\keyword{datasets}
