% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mz_tab.R
\name{writeMzTab}
\alias{writeMzTab}
\title{Write an mzTab tab separated file from the passed in mzTab object.}
\usage{
writeMzTab(mztab, file)
}
\arguments{
\item{mztab}{the R6 MzTab object to write.}

\item{file}{the file path to write to.}
}
\description{
Write an mzTab tab separated file from the passed in mzTab object.
}
