% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzTabWriter.r
\name{writeMzTab}
\alias{writeMzTab}
\title{Write an mzTab tab separated file from the passed in mzTab object.}
\usage{
writeMzTab(mztab, filename)
}
\arguments{
\item{mztab}{the R6 MzTab object to write.}

\item{filename}{the filename to write to.}
}
\description{
Write an mzTab tab separated file from the passed in mzTab object.
}
