% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mz_tab.R
\name{writeMzTabJSON}
\alias{writeMzTabJSON}
\title{Write MzTab object as JSON to file.}
\usage{
writeMzTabJSON(mztab, file)
}
\arguments{
\item{mztab}{the R6 \link{MzTab} object to write.}

\item{file}{the file path to write to.}
}
\description{
Converts the provided \link{MzTab} object to JSON and writes it to the provided file.
}
